==> analyses/05-ModelSelection.Rmd <==
---
title: "07 Model Selection"
author: "Galina M. Jönsson"
date: "20/04/2021"
output: 
  html_document:
    toc: true
    toc_depth: 3
    toc_float: true
---

**To do list:**   

* Fix SummariseRhatDetProb-function to:   
    + Handle lists   
    + handle catLL   
* Detection probability for for catLL   
* Map records    
* Tidy obs vs pred estimate plots      
* Annotate shiny plotting code   
* Generalize first Py summary code chunk   





# Background & standardisation

## Background

#### Here, I compare the three occupancy model formulations:   
- **Model A**: specifies that list length should be considered as a categorical variable. There are 3 classes: BNM records (excluding BMS records), UKBMS records, and NHM records.  
   
- **Model B**: specifies that list length should be considered as both a continuous variable and a categorical variable. There are 3 classes: BNM records (excluding BMS records) are considered continuous, whilst UKBMS records and NHM records are considered as categorical.  
   
- **Model C**: specifies that list length should be considered as both continuous and categorical variables with no year effect estimated for one class (NHM records). There are 3 classes: BNM records (excluding BMS records) are considered continuous, whilst UKBMS records and NHM records are considered as categorical.  



#### For each of 12 species: 
1. *Erebia epiphron*   
2. *Carterocephalus palaemon*   
3. *Melitaea athalia*   
4. *Polyommatus bellargus*   
5. *Thecla betulae*   
6. *Limenitis camilla*   
7. *Polygonia c-album*   
8. *Erebia aethiops*   
9. *Aglais urticae*   
10. *Pararge aegeria*   
11. *Lycaena phlaeas*   
12. *Pieris rapae*   




See the file '04-OccupancyTrends.Rmd' for detailed methods of species selection. In short, species were selected based on range (small, medium or large), weather they are Habitat Specialists or Wider Countryside Species, variance of the mean annual detection probability and record numbers. Species that show much uncertainty (high variance of the mean annual detection probability) are expected to benefit most from the Model C detection sub-model formulation. Note that there are no Small Range Wider Countryside Species nor any Large Range Habitat Specialist; hence, I chose four species from each of Small Range Habitat Specialist and Large Range Wider Countryside Species, respectively.   



**Small Range Habitat Specialist** (four species)   
The three species with the largest mean annual detection probability standard deviations (*Erebia epiphron*, *Carterocephalus palaemon* and *Melitaea athalia*) and the sixth (of 13) most variable species, *Polyommatus bellargus*, which has most records of all species in this group.   
   
   
   
**Medium Range Habitat Specialist** (two species)   
The species with the largest mean annual detection probability standard deviation (*Thecla betulae*) and fourth (of 16) most variable species, *Limenitis camilla*, because among the species in this group, it is the only species with recent range expansions and has among the highest record numbers.   
   
   
   
**Medium Range Wider Countryside Species** (two species)   
The largest mean annual detection probability standard deviation (*Polygonia c-album*) and the third (of 10) most variable species, *Erebia aethiops*, as it has among the fewest records of the species in this group.   
   
   
   
**Large Range Wider Countryside Species** (four species)   
The three species with the largest mean annual detection probability standard deviations (*Aglais urticae*, *Pararge aegeria* and *Lycaena phlaeas*) and the fourth (of 13) most variable species, *Pieris rapae*, as the genus of the third (*Aglais io*) is already represented in the sample.   







```{r load-model-outputs, message=FALSE, warning=FALSE}
require(dplyr)
#require(devtools)
#devtools::install_github("rstudio/rmarkdown")
require(rmarkdown)


# Create a vector with species names to name model outputs
ModelNames <- c('Aglais_urticae',
                'Carterocephalus_palaemon',
                'Erebia_aethiops',
                'Erebia_epiphron',
                'Limenitis_camilla',
                'Lycaena_phlaeas',
                'Melitaea_athalia',
                'Pararge_aegeria',
                'Pieris_rapae',
                'Polygonia_c.album',
                'Polyommatus_bellargus',
                'Thecla_betulae')




########################################################## 
#######################  Model A  ########################
##########################################################

ModelNames_catLL <- paste(ModelNames, "catLL", sep="_") # Add '_catLL' to model names
catLL_list <- list() # Create an empty list
input_dir <- "../outputs/catLL-outputs/"  # Define input directory

# List files to loop through
files <- list.files(path = paste(input_dir), 
                    ignore.case = TRUE, 
                    pattern = '\\.rds$')

# Loop through, read the rds files, append to list and name elements
for(i in 1:length(files)){
  # read rds file and append to list
  catLL_list[[i]] <- readRDS(file.path(input_dir, files[i])) 
  # name element
  names(catLL_list)[[i]] <- paste(ModelNames_catLL[i])
}




########################################################## 
#######################  Model B  ########################
##########################################################

# Load and list model outputs. I specify files as folder contains all species
mixLL_list <- list(
  readRDS("../outputs/mixLL-outputs/results_Aglais_urticae_crick_mixLL.rds"),
  readRDS("../outputs/mixLL-outputs/results_Carterocephalus_palaemon_ctag_mixLL.rds"),
  readRDS("../outputs/mixLL-outputs/results_Erebia_aethiops_crick_mixLL.rds"),
  readRDS("../outputs/mixLL-outputs/results_Erebia_epiphron_crick_mixLL.rds"),
  readRDS("../outputs/mixLL-outputs/results_Limenitis_camilla_ctag_mixLL.rds"),
  readRDS("../outputs/mixLL-outputs/results_Lycaena_phlaeas_crick_mixLL.rds"),
  readRDS("../outputs/mixLL-outputs/results_Melitaea_athalia_watson_mixLL.rds"),
  readRDS("../outputs/mixLL-outputs/results_Pararge_aegeria_watson_mixLL.rds"),
  readRDS("../outputs/mixLL-outputs/results_Pieris_rapae_crick_mixLL.rds"),
  readRDS("../outputs/mixLL-outputs/results_Polygonia_c_album_watson_mixLL.rds"),
  readRDS("../outputs/mixLL-outputs/results_Polyommatus_bellargus_crick_mixLL.rds"),
  readRDS("../outputs/mixLL-outputs/results_Thecla_betulae_crick_mixLL.rds")
)

# Add model name '_mixLL' to species names
ModelNames_mixLL <- paste(ModelNames, "mixLL", sep="_")

# Name list elements accordingly
for(i in 1:length(mixLL_list)){
  names(mixLL_list)[[i]] <- paste(ModelNames_mixLL[i])
}




########################################################## 
#######################  Model C  ########################
##########################################################

ModelNames_mixLL2 <- paste(ModelNames, "mixLL2", sep="_") # Add '_mixLL2' to model names
mixLL2_list <- list() # Create an empty list
input_dir <- "../outputs/mixLL2-outputs/" # Define input directory

# List files to loop through
files <- list.files(path = paste(input_dir), 
                    ignore.case = TRUE, 
                    pattern = '\\.rds$')

# Loop through, read the rds files, append to list and name elements
for(i in 1:length(files)){
  # read rds file and append to list
  mixLL2_list[[i]] <- readRDS(file.path(input_dir, files[i])) 
  # name element
  names(mixLL2_list)[[i]] <- paste(ModelNames_mixLL2[i])
}




########################################################## 
#############  Check model A, B & C outputs ##############
##########################################################

# Check that the number of species observations and sites where the species have been collected and recorded are equal for model B and C. In other words whether the models are based on the same data set. 
for (i in 1:length(mixLL_list)) {
  if(isFALSE(mixLL_list[[i]]$species_sites == mixLL2_list[[i]]$species_sites &&
             catLL_list[[i]]$species_sites == mixLL_list[[i]]$species_sites &&
             catLL_list[[i]]$species_sites == mixLL2_list[[i]]$species_sites)){
    print("stop, species_observations differ")
  }
  if(isFALSE(mixLL_list[[i]]$species_sites == mixLL2_list[[i]]$species_sites &&
             catLL_list[[i]]$species_sites == mixLL_list[[i]]$species_sites &&
             catLL_list[[i]]$species_sites == mixLL2_list[[i]]$species_sites)){
    print("stop, species_sites differ")
  }
}




########################################################## 
##################  Rhat: Convergence ####################
##########################################################


### occupancy estimates

# Source function checking occupancy convergence i.e. Rhat >1.1?
source("./function-summarise_Rhat.R")

# If some species have not converged (converged-column == "N"), subset them
NotConvergedOccu <- subset((summarise_Rhat(append(catLL_list, 
                                              (append(mixLL_list,
                                              mixLL2_list))), 
                                       verbose=FALSE,
                                       is.list=TRUE)), # Specify its a list
                          grepl("N", converged))


# Source function checking detection probability convergence i.e. Rhat >1.1?
source("./function-SummariseRhatDetProb.R")

# Run function
#RhatDet_catLL <- SummariseRhatDetProb("../outputs/catLL-outputs/", 
#                              verbose=FALSE)


# If some species have not converged (converged-column == "N"), subset them
#NotConvergedDetProb_catLL <- subset(RhatDet_mixLL, grepl("N", RhatDet_catLL[,2:5]))
NotConvergedDetProb_mixLL <- subset((SummariseRhatDetProb("../outputs/mixLL-outputs/", 
                              verbose=FALSE)), 
                              grepl("N", (SummariseRhatDetProb("../outputs/mixLL-outputs/", 
                              verbose=FALSE))[,2:5]))
NotConvergedDetProb_mixLL2 <- subset((SummariseRhatDetProb("../outputs/mixLL2-outputs/", 
                              verbose=FALSE)), 
                              grepl("N", (SummariseRhatDetProb("../outputs/mixLL2-outputs/", 
                              verbose=FALSE))[,2:5]))


# If any of the model outputs have occupancy estimates that have not converged (Rhat >1.1), display them as a table.
if(nrow(NotConvergedOccu) > 0){
  knitr::kable(NotConvergedOccu)
}
if(nrow(NotConvergedDetProb_mixLL) > 0){
  knitr::kable(NotConvergedDetProb_mixLL)
}
if(nrow(NotConvergedDetProb_mixLL2) > 0){
  knitr::kable(NotConvergedDetProb_mixLL2)
}
```


## Visualisation




### Record numbers
```{r plot-record-numbers, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE}
# Load required packages
require(dplyr)
require(plyr)
require(ggplot2)
require(viridis)
require(readr)

# Load spp_vis data (any would do), which contain all visits and whether or not each of the three species have been observed at that visit
spp_vis <- read.csv('../data/formattedData/spp_vis_merged_2020-04-07.csv', header=T, na.strings=c("","NA"))

# Load occDetdata giving each visits' site ID (grid cell), list length (numper of species observed) and time period (in this case year)
occDetdata_merged_2020_04_07 <- read.csv('../data/formattedData/occDetdata_merged_2020-04-07.csv', header=T, na.strings=c("","NA"))

# Merge by visit
spp_vis <- merge(spp_vis, occDetdata_merged_2020_04_07, by="visit")

# Find unique visits
spp_vis <- unique(spp_vis)

# Create a species column
spp_vis$species <- as.factor(c("NA"))

### Make L column a source column
# Change all List lengths between 0 and 22222 to 1 (all these are BNM data)
spp_vis$L <- replace(spp_vis$L, spp_vis$L > 0 & spp_vis$L < 22222, 1)
# Make L a factor
spp_vis$L <- as.factor(spp_vis$L)
# Rename factor levels according to data type
spp_vis$L <- revalue(spp_vis$L, c("1"="BNM", "22222"="UKBMS", "33333"="NHCs"))
# Rename column to Source
colnames(spp_vis)[colnames(spp_vis)=="L"] <- "Source"


# Correct time period so it starts at 1900
spp_vis$year <- spp_vis$TP+1899

### Find visits per data type
#test <- spp_vis %>% group_by(Source) %>% tally()
#test

### Find records per data type
#spp_vis$count.TRUE <- apply(spp_vis, 1, function(x) length(which(x=="TRUE")))
#



########################################################################

# Subset each species' records into dataframes and specify species identity
spp_vis_A <- subset(spp_vis, Aglais.urticae == "TRUE")
spp_vis_A$species <- c("italic('A. urticae')")
spp_vis_B <- subset(spp_vis, Pararge.aegeria == "TRUE")
spp_vis_B$species <- c("italic('P. aegeria')")
spp_vis_C <- subset(spp_vis, Polygonia.c.album == "TRUE")
spp_vis_C$species <- c("italic('P. c-album')")
spp_vis_D <- subset(spp_vis, Pieris.rapae == "TRUE")
spp_vis_D$species <- c("italic('P. rapae')")


# row bind each subsetted species' dataframe with the original spp_vis dataframe
spp_vis1 <- rbind(spp_vis_A, spp_vis_B)
spp_vis1 <- rbind(spp_vis1, spp_vis_C)
spp_vis1 <- rbind(spp_vis1, spp_vis_D)




########################################################################


# Subset each species' records into dataframes and specify species identity
spp_vis_Carterocephalus_palaemon <- subset(spp_vis, Carterocephalus.palaemon == "TRUE")
spp_vis_Carterocephalus_palaemon$species <- c("italic('C. palaemon')")
spp_vis_Erebia_aethiops <- subset(spp_vis, Erebia.aethiops == "TRUE")
spp_vis_Erebia_aethiops$species <- c("italic('E. aethiops')")
spp_vis_Erebia_epiphron <- subset(spp_vis, Erebia.epiphron == "TRUE")
spp_vis_Erebia_epiphron$species <- c("italic('E. epiphron')")
spp_vis_Melitaea_athalia <- subset(spp_vis, Melitaea.athalia == "TRUE")
spp_vis_Melitaea_athalia$species <- c("italic('M. athalia')")



# row bind each subsetted species' dataframe with the original spp_vis dataframe
spp_vis2 <- rbind(spp_vis_Carterocephalus_palaemon, spp_vis_Erebia_aethiops)
spp_vis2 <- rbind(spp_vis2, spp_vis_Melitaea_athalia)
spp_vis2 <- rbind(spp_vis2, spp_vis_Erebia_epiphron)




########################################################################


# Subset each species' records into dataframes and specify species identity
spp_vis_Limenitis_camilla <- subset(spp_vis, Limenitis.camilla == "TRUE")
spp_vis_Limenitis_camilla$species <- c("italic('L. camilla')")
spp_vis_Lycaena_phlaeas <- subset(spp_vis, Lycaena.phlaeas == "TRUE")
spp_vis_Lycaena_phlaeas$species <- c("italic('L. phlaeas')")
spp_vis_Polyommatus_bellargus <- subset(spp_vis, Polyommatus.bellargus == "TRUE")
spp_vis_Polyommatus_bellargus$species <- c("italic('P. bellargus')")
spp_vis_Thecla_betulae <- subset(spp_vis, Thecla.betulae == "TRUE")
spp_vis_Thecla_betulae$species <- c("italic('T. betulae')")


# row bind each subsetted species' dataframe with the original spp_vis dataframe
spp_vis3 <- rbind(spp_vis_Limenitis_camilla, spp_vis_Lycaena_phlaeas)
spp_vis3 <- rbind(spp_vis3, spp_vis_Polyommatus_bellargus)
spp_vis3 <- rbind(spp_vis3, spp_vis_Thecla_betulae)


################################################################

# Plot function
plot_speciesVisits <- function(data, title=NULL){
pp <- ggplot(data, aes(x=year, fill=Source, color=Source)) + 
  scale_fill_manual(values = alpha(c("#453781FF", 
                                     "#20A386FF", 
                                     "#DCE318FF"), 0.5)) +
  scale_color_manual(values = alpha(c("#453781FF", 
                                      "#20A386FF", 
                                      "#DCE318FF"), 0.5)) +
  facet_grid(~ species, labeller=label_parsed) + # facet grid by species
  geom_bar(position="identity", alpha=0.7) + 
  labs(y="Number of Visits", x = "Year") +
  theme_bw() +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank()) +
  # sqrt y scale and expand scale to no space below, 5% above
  scale_y_sqrt(expand = expansion(mult = c(0, .05))) +
  theme(text = element_text(size=15),
        axis.text.x = element_text(angle=90, hjust=1))
print(pp)
}
  

# Apply plot function to the three species lists
mapply(data = list(spp_vis1, 
                  spp_vis3, 
                  spp_vis2),
       FUN = plot_speciesVisits)

```







### DIC

#### 
Plot 
Use model C output as default & plot \Delta$ DIC for C:A & C:B. 
```{r DIC-delta, cache=TRUE, message=FALSE, echo=FALSE}
deltaDIC_out <- data.frame(Species = as.character(),
                           Model_A = as.numeric(),
                           Model_B = as.numeric(),
                           Model_C = as.numeric(),
                           Delta_AC = as.numeric(),
                           Delta_BC = as.numeric())

# loop through BUGSoutput for each species pair in turn
for (i in 1:length(catLL_list)){ 
  deltaDIC_out[i, "Species"] <- paste(paste("*", ModelNames[i], sep = ""), "*", sep = "")     # Populate species_name column from spp_names list and surround by asterisks
  deltaDIC_out[i, "Model_A"] <-  catLL_list[[i]]$BUGSoutput$DIC     # Enter Model A's DIC value
  deltaDIC_out[i, "Model_B"] <-  mixLL_list[[i]]$BUGSoutput$DIC     # Enter Model B's DIC value
  deltaDIC_out[i, "Model_C"] <-  mixLL2_list[[i]]$BUGSoutput$DIC     # Enter Model C's DIC value
  deltaDIC_out[i, "Delta_AC"] <- mixLL2_list[[i]]$BUGSoutput$DIC - catLL_list[[i]]$BUGSoutput$DIC
  deltaDIC_out[i, "Delta_BC"] <- mixLL2_list[[i]]$BUGSoutput$DIC - mixLL_list[[i]]$BUGSoutput$DIC
}

deltaDIC_out2 <- data.frame(Species = as.character(),
                           Model = as.character(),
                           Delta = as.numeric())

deltaDIC_out2[1:24, "Species"] <- deltaDIC_out$Species
deltaDIC_out2[1:12, "Model"] <- "C - A"
deltaDIC_out2[1:12, "Delta"] <- deltaDIC_out$Delta_AC
deltaDIC_out2[13:24, "Model"] <- "C - B"
deltaDIC_out2[13:24, "Delta"] <- deltaDIC_out$Delta_BC



deltaDIC_out2[12, "Delta"] <- -87.86348

library(ggplot2)


# Grouped
ggplot(deltaDIC_out2, aes(fill=Model, y=Delta, x=Species)) + 
  geom_bar(position="dodge", stat="identity") + ylab("Delta DIC") +
    geom_hline(yintercept = 0) +
 # scale_y_continuous(trans='log10') +
  theme(panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank(),
        panel.background = element_blank(), 
        axis.line = element_line(colour = "black")) +
  coord_flip()



#scale_x_continuous(trans=‘log2’)


knitr::kable(deltaDIC_out)
```
A more informative way to 



```{r DIC, cache=TRUE, message=FALSE, echo=FALSE}
# Create empty data frame to populate
summary_out <- data.frame(Species = as.character(),
                          Best_fit = as.character(),
                          Model_B = as.numeric(),
                          Model_C = as.numeric(),
                          Difference = as.numeric())

# Change ModelNames vetor to species names with spaces
spp_names <- gsub(ModelNames, patt = "_", repl=" ")
spp_names <- sub(spp_names, patt = "c.album", repl="c-album")

# loop through BUGSoutput for each species pair in turn
for (i in 1:length(mixLL_list)) { # If Model B's DIC is lower than C's
  if(mixLL_list[[i]]$BUGSoutput$DIC < mixLL2_list[[i]]$BUGSoutput$DIC) {
    Best <- "B"    # Name 'Best' B
  } else {
      Best <- "C"    # Else name 'Best' C
  }
  summary_out[i, "Species"] <- paste(paste("*", spp_names[i], sep = ""), "*", sep = "")     # Populate species_name column from spp_names list and surround by asterisks
  summary_out[i, "Best_fit"] <-  Best     # Populate Best_fit column with lowest DIC model
  summary_out[i, "Model_B"] <-  mixLL_list[[i]]$BUGSoutput$DIC     # Enter Model B's DIC value
  summary_out[i, "Model_C"] <-  mixLL2_list[[i]]$BUGSoutput$DIC     # Enter Model C's DIC value
  summary_out[i, "Difference"] <- abs(mixLL_list[[i]]$BUGSoutput$DIC - mixLL2_list[[i]]$BUGSoutput$DIC)     # Enter the difference between Model B's and C's DIC values
  }

knitr::kable(summary_out)
```
**Reminder**: DIC differences are 'absolute', not relative to the DIC estimates themselves (e.g. a difference of 2 is the same thing when models' DICs are 80k and 79998 vs 2 and 4?)


# Posterior Predictive Checks

## Extract and summarise Py[j] posterior distribution

> (1)   
For each of *v* visits in each species’ model, extract 99 samples from the posterior distribution of the probability that an observation was made on that visit. In practical terms, this probability is the product of the true (unknown) occupancy, *z*~*it*~, and the detection probability, *p*~*itv*~


As the model formulation is coded in the following way:   
```{r model-example, cache=TRUE, message=FALSE, eval=FALSE}
for(j in 1:nvisit) {   
y[j] ~ dbern(Py[j])   
y.new[j] ~ dbern(Py[j])   
Py[j] <- z[Site[j],Year[j]]*p[j]   
logit(p[j]) <-  alpha.p[Year[j]] + LL.p * logL[j] + dtype2.p * DATATYPE2[j] + dtype3.p * DATATYPE3[j] }   
}  
```
   
   
the following line gives the product of the true occupancy, *z*~*it*~, and the detection probability, *p*~*itv*~
```{r py-demonstration, cache=TRUE, message=FALSE, eval=FALSE}
Py[j] <- z[Site[j],Year[j]]*p[j]   
```   

In other words, we want to extract Py[j]

> (2)   
Use each of the 99 sets of probabilities to sample a vector of potential observations under the model, by treating each visit as a potential Bernoulli trial. These vectors are 99 realizations of *y*~rep~

> (3)   
For each realization *y*~rep~ and for each year, calculate the annual proportion of sites in which the species was recorded; denote this proportion, for species *s* and year *t*, by *T*~*st*~(*y*~rep~)


```{r PostPredChecks1, eval=FALSE, message=FALSE}
require("rjags")
require("coda")
require("reshape2")
require("dplyr")
require("sparta")
require("lattice")
require("LearnBayes")
require("R2jags")


Thecla_betulae.catLL <- catLL_list[[12]]


###########################
############ 1 ############
###########################


# recompile the model
Thecla_betulae.catLL$model$recompile()

#out$model$recompile()

# Sample 100 iterations of each of the 3 chains and thin by 3, giving 99 samples
samp <- rjags:::coda.samples(model=Thecla_betulae.catLL$model, 
                             parallel=TRUE, n.cores=3,
                             variable.names="Py", 
                             n.iter=100, # Define number of iterations to run
                             thin=3) # Define thinning


###########################
############ 2 ############
###########################

### Create vectors that are are 99 realizations of yrep

# Use the 99 sets of probabilities & sample a vector of simulated observations under the model, by treating each visit as a potential Bernoulli trial. Here, we use rbinom as Bernoulli trial 
Y_rep <- lapply(samp, function(x) apply(x, 1:2, rbinom, n=1, size=1))


### Tidy the 99 realizations of yrep       

# What does this do? 
Y_rep <- melt(Y_rep)

# Tidy VisitID
Y_rep$VisitID <- as.character(gsub(Y_rep$Var2, pa="Py\\[", repl=""))
Y_rep$VisitID <- as.numeric(gsub(Y_rep$VisitID, pa="\\]", repl=""))

# Tidy iteration
Y_rep$iter <- with(Y_rep, Var1 + max(Var1)*(L1 - 1))

# Include year, siteID and the actual observation (y)        
# load the raw data
#temp_data <- out$model$data()
temp_data <- Thecla_betulae.catLL$model$data()
temp_data <- as.data.frame(with(temp_data, cbind(Site, Year, y)))
temp_data$VisitID <- 1:nrow(temp_data)
Y_rep <- merge(Y_rep[,-c(1,2,4)], temp_data)
        
#with(Y_rep, table(y,value))


###########################
############ 3 ############
###########################
        
# next calculate the reporting rate in differing ways
# first, for each site:year combination (i.e., yrep) report the number of sites for which a positive record was made
RR1 <- Y_rep %>% group_by(Site, Year, iter) %>%
  summarise(obs = max(y),sim = max(value)) %>%
  ungroup()
# the data in RR1 refer to whether the species was recorded in each site in each year
# save this file
write.csv(RR1, "../outputs/posteriorPredictiveChecks-outputs/Thecla_betulae.catLL_RR1.csv")

# next, the mean of these across years (i.e., Tst(yrep))
RR2 <- RR1 %>% group_by(Year, iter) %>%
          summarise(obs_meanSitesPerYear = mean(obs),
                    sim_meanSitesPerYear = mean(sim)
          ) %>%
          ungroup()

# the data in RR2 refer to the proportion of Sites with positive records per Year
# save this file
write.csv(RR2, "../outputs/posteriorPredictiveChecks-outputs/Thecla_betulae.catLL_RR2.csv")
```





> (4)
Calculate *T*~*gt*~(*y*~rep~) for each replicate dataset as the mean of *T*~*st*~(*y*~rep~) across species in each taxonomic group, *g*.

Here, I will calculate the mean of *T*~*st*~(*y*~rep~) across species for each model formulation, not taxonomic group (still it calling *g*).


> (5)
Calculate the mean *m* across years as *T*~*gm*~(*y*~rep~) for each replicate dataset


```{r read-PostPredCheck-Data, message=FALSE, warning=FALSE}
library(R2jags)
library(reshape2)
library(sparta)
library(LearnBayes)
library(lattice)
library(dplyr)
library(tidyr)
library(ggplot2)

# Define 99% CIs
q <- c(0.005, 0.995)

# Specify directory
path_in <- "../outputs/posteriorPredictiveChecks-outputs/"

# List relevant files in directory
spp_files <- paste(path_in, list.files(path=path_in, 
                                       pattern="\\RR2.csv$",
                                       recursive = TRUE), sep="/")
# read relevant files
data <- lapply(spp_files, read.csv)

# how many models?
print(length(data))

#append the species names
sppnames <- gsub(spp_files, patt = paste0(path_in, "/"), repl="")
sppnames <- gsub(sppnames, patt = "_RR2.csv", repl="")
sppnames <- sub(sppnames, patt = "_", repl=".")
names(data) <- sppnames


# coerce to a single dataframe
sppMeanSites <- melt(data, id=1:5)

# separate Model and Species
sppMeanSites <- sppMeanSites %>% 
  separate(L1, into=c("Species", "Model"), 
           sep="_", remove=TRUE)

# Assign correct years
sppMeanSites$Year <- sppMeanSites$Year + 1899

# Correct species names and include asteriks for italising 
sppMeanSites$Species <- gsub(sppMeanSites$Species, patt = "c.album", repl="c-album")
sppMeanSites$Species <- gsub(sppMeanSites$Species, patt = "[.]", repl=" ")
sppMeanSites$Species <- paste("*", sppMeanSites$Species, sep = "")
sppMeanSites$Species <- paste(sppMeanSites$Species, "*", sep = "")
```


Now we can calculate three other summary statistics:
* sppVarSites: the variance across years in the proportion of sites with a record
* modelMeanSites: the annual mean (across species) in the proportion of sites with a record per model
* modelVarSites: the variance across years in the above



> (6)
Calculate *T*~*gvar*~(*y*~*rep*~) as the variance across years in *T*~*gt*~(*y*~rep~) for each replicate dataset.

> (7)
Calculate the observed mean proportion of sites with records, *T*~*gm*~(*y*), and the variance across years, *T*~*gvar*~(*y*), for each replicate dataset.


```{r varianceAndModel, message=FALSE, warning=FALSE}
# now the variance across years 
sppVarSites <- sppMeanSites %>% dplyr:::group_by(Model, Species, iter) %>%
  dplyr:::summarise(obs_varSitesPerYear = var(obs_meanSitesPerYear),
            sim_varSitesPerYear = var(sim_meanSitesPerYear)
  ) %>%
  dplyr:::ungroup()

# now summarise these stats to model level
# We'll calculate the mean proportion of sites with records for species in each taxon
modelMeanSites <- sppMeanSites %>% dplyr:::group_by(Model, Year, iter) %>%
  dplyr:::summarise(obs_meanSitesPerYear = mean(obs_meanSitesPerYear),
            sim_meanSitesPerYear = mean(sim_meanSitesPerYear)
  ) %>%
  dplyr:::ungroup()

# finally, the variance among years
modelVarSites <- modelMeanSites %>% dplyr:::group_by(Model, iter) %>%
  dplyr:::summarise(obs_varSitesPerYear = var(obs_meanSitesPerYear),
            sim_varSitesPerYear = var(sim_meanSitesPerYear)
  ) %>%
  ungroup()


#####################################
sppMeanSitesYear <- sppMeanSites %>% dplyr:::group_by(Model, Species, Year) %>%
  dplyr:::summarise(obs_meanSitesPerYear = mean(obs_meanSitesPerYear),
            sim_MeanSitesPerYear = mean(sim_meanSitesPerYear),
            sim_lower = quantile(sim_meanSitesPerYear, q[1]),
            sim_upper = quantile(sim_meanSitesPerYear, q[2])
  ) %>%
  dplyr:::ungroup()

sppMeanSitesYear$Species <- as.factor(sppMeanSitesYear$Species)
```


> (8)
Summarize the distributions of *T*~*gm*~(*y*) and *T*~*gvar*~(*y*) as the mean and 95% credible intervals to demonstrate the variation in summary measures that can reasonably be expected under the model.

I will use 99% CIs as the lower 95% CI overlaps zero. 

Summarise the data, including the Bayesian *p* value

```{r BayesianP, message=FALSE, warning=FALSE}
sppMeanSites_Summ <- sppMeanSites %>% 
  dplyr:::group_by(Model, Species) %>%
    dplyr:::summarise(Bp = mean(obs_meanSitesPerYear > sim_meanSitesPerYear),
              obs = mean(obs_meanSitesPerYear),
              sim_mean = mean(sim_meanSitesPerYear),
              sim_lower = quantile(sim_meanSitesPerYear, q[1]),
              sim_upper = quantile(sim_meanSitesPerYear, q[2]),
              ) %>%
  dplyr:::ungroup()


sppVarSites_Summ <- sppVarSites %>% 
  dplyr:::group_by(Model, Species) %>%
  dplyr:::summarise(Bp = mean(obs_varSitesPerYear > sim_varSitesPerYear),
            obs = mean(obs_varSitesPerYear),
            sim_mean = mean(sim_varSitesPerYear),
            sim_lower = quantile(sim_varSitesPerYear, q[1]),
            sim_upper = quantile(sim_varSitesPerYear, q[2]),
  ) %>%
  dplyr:::ungroup()


modelMeanSites_Summ <- modelMeanSites %>% 
  dplyr:::group_by(Model) %>%
  dplyr:::summarise(Bp = mean(obs_meanSitesPerYear > sim_meanSitesPerYear),
            obs = mean(obs_meanSitesPerYear),
            sim_mean = mean(sim_meanSitesPerYear),
            sim_lower = quantile(sim_meanSitesPerYear, q[1]),
            sim_upper = quantile(sim_meanSitesPerYear, q[2]),
  ) %>%
  dplyr:::ungroup()


modelVarSites_Summ <- modelVarSites %>% 
  dplyr:::group_by(Model) %>%
  dplyr:::summarise(Bp = mean(obs_varSitesPerYear > sim_varSitesPerYear),
            obs = mean(obs_varSitesPerYear),
            sim_mean = mean(sim_varSitesPerYear),
            sim_lower = quantile(sim_varSitesPerYear, q[1]),
            sim_upper = quantile(sim_varSitesPerYear, q[2]),
  ) %>%
  dplyr:::ungroup()

```


### Visualise Py summary {.tabset .tabset-dropdown}

#### Observed vs Predicted data
```{r plot-obsVsPred, echo=FALSE, message=FALSE, warning=FALSE, cache=TRUE}
require(ggtext)
require(cowplot)

modelMeanBp_Summ <- modelMeanSites %>% 
  dplyr:::group_by(Model) %>%
  dplyr:::summarise(Bp = mean(obs_meanSitesPerYear > sim_meanSitesPerYear),
            obs = mean(obs_meanSitesPerYear),
            sim_mean = mean(sim_meanSitesPerYear),
            sim_lower = quantile(sim_meanSitesPerYear, q[1]),
            sim_upper = quantile(sim_meanSitesPerYear, q[2]),
  ) %>%
  dplyr:::ungroup()

plot_PredPostTest <- function(data, title=NULL){
  gp <- ggplot(data = data,) +
    geom_point(aes(x=obs, y=sim_mean, shape=Species, col=Model), size=3.5, alpha=0.5) +
    geom_errorbar(aes(x=obs, ymin=sim_lower, ymax=sim_upper, col=Model), 
               alpha=0.5, size = 0.8) +
    scale_shape_manual(values = c(1,2,3,4,5,6,60,8,18,17,16,15)) +
    scale_colour_manual(values =c("#440154FF", "#20A387FF", "#E69F00")) +
    geom_abline(aes(intercept=0, slope=1)) + 
    scale_x_log10() + scale_y_log10() + 
    xlab("Observed from data") + ylab("Model prediction") + 
    ggtitle(title) + theme_bw() +
    theme(panel.grid.minor = element_blank()) +
    theme(text=element_text(size=13)) + 
    theme(legend.text = element_markdown())
  print(gp)
}


########################################
### Re-plot for thesis Figure 2.5A-B
########################################

plot_PredPostTest(sppMeanSites_Summ)
#ggsave("../figs/Fig2-5A.png")
plot_PredPostTest(sppVarSites_Summ)
#ggsave("../figs/Fig2-5B.png")

#ggsave("../figs/Fig2-5-legend.png", get_legend(plot_PredPostTest(sppVarSites_Summ)))
  

#mapply(data = list(sppMeanSites_Summ, 
#                  sppVarSites_Summ), # 
                  #modelMeanSites_Summ,
                  #modelVarSites_Summ), 
#        title = list("Mean (across years) proportion of sites with a record - 95%CIs", 
#                  "Variance (across years) in proportion of sites per year - 95%CIs"), #, 
                  #"Mean proportion of sites per year (mean across species) - 99%CIs",
                  #"Variance across years in proportion of sites per year across species - 99%CIs"),   
#       FUN = plot_PredPostTest)
```




#### Bayesian *p*: Model A vs Model B


Now plot the Bayesian *p*-values (the proportion of simulated values that are lower than the observed; the 'best' Bayesian *p*-value is 0.5) 
```{r plot-Bp-AvsB, echo=FALSE, warning=FALSE, cache=TRUE}
require(tidyr)
require(viridis)
require(ggtext)

ModelMeanBp_Summ <- spread((sppMeanSites_Summ[,1:3]), Model, Bp)
ModelVarBp_Summ <- spread((sppVarSites_Summ[,1:3]), Model, Bp)

# Rename
names(ModelVarBp_Summ)[2] <- "catLL_var"
names(ModelVarBp_Summ)[3] <- "mixLL_var"
names(ModelVarBp_Summ)[4] <- "mixLL2_var"

ModelBp_Summ <- data.frame(ModelMeanBp_Summ)
ModelVarBp_Summ <- data.frame(ModelVarBp_Summ)
ModelBp_Summ[,5:7] <- data.frame(ModelVarBp_Summ[,2:4])


ggplot(ModelBp_Summ, aes(x=catLL, y=mixLL, group=Species)) + 
  geom_point(aes(shape=Species, color=Species),  size=5) +
  scale_color_manual(values =c (viridis(15))) +
  scale_shape_manual(values = c(15:18, 21:24, 15:18, 21:23)) + 
  ggtitle("Bayesian p-values",
          subtitle = "Mean proportion of sites with records across years") + 
  theme_bw() +
  geom_abline(aes(intercept=0, slope=1)) + 
  xlab("Model A") + ylab("Model B") +
  geom_hline(yintercept=0.5, linetype="dashed", color = "red", alpha=0.5) +
  geom_vline(xintercept=0.5, linetype="dashed", color = "red", alpha=0.5) +
  theme(legend.text = element_markdown()) +
  theme(text=element_text(size=13)) 


ggplot(ModelBp_Summ, aes(x=catLL_var, y=mixLL2_var, group=Species)) +
  geom_point(aes(shape=Species, color=Species),  size=5) +
  scale_color_manual(values =c(viridis(15))) +
  scale_shape_manual(values = c(15:18, 21:24, 15:18, 21:23)) + 
  ggtitle("Bayesian p",
          subtitle = "Variance proportion of sites with records across years") + 
  theme_bw() +
  geom_abline(aes(intercept=0, slope=1)) + 
  xlab("Model A") + ylab("Model B") +
  geom_hline(yintercept=0.5, linetype="dashed", color = "red", alpha=0.5) +
  geom_vline(xintercept=0.5, linetype="dashed", color = "red", alpha=0.5) +
  theme(legend.text = element_markdown()) +
  theme(text=element_text(size=13)) 
```
  


#### Bayesian *p*: Model A vs. Model C
```{r plot-Bp-AvsC, echo=FALSE, warning=FALSE, cache=TRUE}
require(tidyr)
require(viridis)
require(ggtext)

ggplot(ModelBp_Summ, aes(x=catLL, y=mixLL2, group=Species)) + 
  geom_point(aes(shape=Species, color=Species),  size=5) +
  scale_color_manual(values =c(viridis(15))) +
  scale_shape_manual(values = c(15:18, 21:24, 15:18, 21:23)) + 
  ggtitle("Bayesian p-values",
          subtitle = "Mean proportion of sites with records across years") + 
  theme_bw() +
  geom_abline(aes(intercept=0, slope=1)) + 
  xlab("Model A") + ylab("Model C") +
  geom_hline(yintercept=0.5, linetype="dashed", color = "red", alpha=0.5) +
  geom_vline(xintercept=0.5, linetype="dashed", color = "red", alpha=0.5) +
  theme(legend.text = element_markdown()) +
  theme(text=element_text(size=13)) 

ggplot(ModelBp_Summ, aes(x=catLL_var, y=mixLL2_var, group=Species)) +
  geom_point(aes(shape=Species, color=Species),  size=5) +
  scale_color_manual(values =c(viridis(15))) +
  scale_shape_manual(values = c(15:18, 21:24, 15:18, 21:23)) + 
  ggtitle("Bayesian p",
          subtitle = "Variance proportion of sites with records across years") + 
  theme_bw() +
  geom_abline(aes(intercept=0, slope=1)) + 
  xlab("Model A") + ylab("Model C") +
  geom_hline(yintercept=0.5, linetype="dashed", color = "red", alpha=0.5) +
  geom_vline(xintercept=0.5, linetype="dashed", color = "red", alpha=0.5) +
  theme(legend.text = element_markdown()) +
  theme(text=element_text(size=13)) 
```


#### Bayesian *p*: Model B vs. Model C
```{r plot-Bp-BvsC, echo=FALSE, warning=FALSE, cache=TRUE}
require(tidyr)
require(viridis)
require(ggtext)

ggplot(ModelBp_Summ, aes(x=mixLL, y=mixLL2, group=Species)) + 
  geom_point(aes(shape=Species, color=Species),  size=5) +
  scale_color_manual(values =c(viridis(15))) +
  scale_shape_manual(values = c(15:18, 21:24, 15:18, 21:23)) + 
  ggtitle("Bayesian p-values",
          subtitle = "Mean proportion of sites with records across years") + 
  theme_bw() +
  geom_abline(aes(intercept=0, slope=1)) + 
  xlab("Model B") + ylab("Model C") +
  geom_hline(yintercept=0.5, linetype="dashed", color = "red", alpha=0.5) +
  geom_vline(xintercept=0.5, linetype="dashed", color = "red", alpha=0.5) +
  theme(legend.text = element_markdown()) +
  theme(text=element_text(size=13)) 


ggplot(ModelBp_Summ, aes(x=mixLL_var, y=mixLL2_var, group=Species)) +
  geom_point(aes(shape=Species, color=Species),  size=5) +
  scale_color_manual(values =c(viridis(15))) +
  scale_shape_manual(values = c(15:18, 21:24, 15:18, 21:23)) + 
  ggtitle("Bayesian p",
          subtitle = "Variance proportion of sites with records across years") + 
  theme_bw() +
  geom_abline(aes(intercept=0, slope=1)) + 
  xlab("Model B") + ylab("Model C") +
  geom_hline(yintercept=0.5, linetype="dashed", color = "red", alpha=0.5) +
  geom_vline(xintercept=0.5, linetype="dashed", color = "red", alpha=0.5) +
  theme(legend.text = element_markdown()) +
  theme(text=element_text(size=13)) 
```
**Notes**   

* Model B/mixLL: Six of 12 species have better p-values   
* Model C/mixLL2: Six of 12 species have better p-values   
* DIC-values favor Model B for all species but three: *M athalia*, *C. palaemon* (Model C better according to p-values) and *L. phlaeas* (p-values practically identical)    
* Two species have practically equal p-value:  
    + *L. phlaeas*: some differences in occupancy trends but this species shows cyclical    occupancies over time and both models capture the cyclical pattern     
    + *P. aegeria*: Practically identical occupancy trends     
* *Erebia epiphron*  shows the largest difference in p-vale between the two models    
    + Occupancy estimates for the two models are practically identical       
    + This species has fewest BMS (i.e. standardized) records       
    + This species has fewest NHC records       
* *Limenitis camilla*  shows the second largest difference in p-vale between the two models (0.48 for mixLL2/Model C and 0.57 for for mixLL/Model B)     
    + Occupancy estimates for the two models are practically identical (although much underestimated in model A)       


**Means Bayesian-p: Model B vs C**

1. *Erebia epiphron*: C much better     
2. *Carterocephalus palaemon*: C better    
3. *Melitaea athalia*: C better    
4. *Polyommatus bellargus*: B better    
5. *Thecla betulae*: B slightly better    
6. *Limenitis camilla*: C much better    
7. *Polygonia c-album*: B better    
8. *Erebia aethiops*: C better    
9. *Aglais urticae*: B better    
10. *Pararge aegeria*: practically identical    
11. *Lycaena phlaeas*: practically identical     
12. *Pieris rapae*: B slightly better      


**Nick**   
   
* Given that the 'optimal' p-value is 0.5, is *Limenitis camilla* (0.48 for mixLL2/Model C and 0.57 for for mixLL/Model B) "more" interesting than let's say *Erebia epiphron* (which shows the largest difference in p-vale between the two models)?    Two questions: which model is better ( for most id doesnt make a difference but for E. epophron)
* Variance: **very** poor p-values...
* Is the sum of of distance from 0.5 across species for model B (1.597614) vs model C (1.555207) of interest?  
* Data visualisation: colour blind-friendly vs. visually appealing... 



```{r BpDiff, echo=FALSE, warning=FALSE, cache=TRUE, eval=FALSE}
mixLL_out <- 0
mixLL2_out <- 0
for(i in 1:nrow(ModelBp_Summ)){
  mixLLtemp <- abs(0.5 - ModelBp_Summ[i,"mixLL"])
  mixLL2temp <- abs(0.5 - ModelBp_Summ[i,"mixLL2"])
  mixLL_out <- mixLL_out + mixLLtemp
  mixLL2_out <- mixLL2_out + mixLL2temp
}

for(i in 1:nrow(ModelBp_Summ)){
  mixLL2temp <- abs(0.5 - ModelBp_Summ[i,"mixLL2"])
}
```
  
  
  
```{r plotBp2, echo=FALSE, warning=FALSE, eval=FALSE}
hist(sppMeanSites_Summ$Bp)
hist(sppVarSites_Summ$Bp)
hist(modelMeanSites_Summ$Bp) ### MixLL_2 on the left
hist(modelVarSites_Summ$Bp)
```


```{r savePySummary, echo=FALSE}
# Finally save the outputs
#write.csv(taxonMeanSites_Summ, file="taxonMeanSites_Summ.csv")
#write.csv(taxonVarSites_Summ, file="taxonVarSites_Summ.csv")
```


# Root-Mean-Square Error (RMSE)
$$RMSE = \sqrt{\sum_{i=1}^{N} \frac{(Predicted_{i}-Actual_{i})^2}{N}}$$


```{r RMSE, echo=FALSE}
# Importing the required package
library(Metrics)
  
# Taking two vectors
#actual = c(1.5, 1.0, 2.0, 7.4, 5.8, 6.6)         
#predicted = c(1.0, 1.1, 2.5, 7.3, 6.0, 6.2)      
  
# Calculating RMSE using rmse()         
#result = rmse(actual, predicted)
  
# Printing the value
#print(result)
```




# Species-wise summary

```{r plot-occu-detProb-setup, message=FALSE, warning=FALSE, echo=FALSE}

### Before plotting, we format all models such that the species-wise plot code is short and sweet


require(ggplot2)
require(labelmachine)
require(sparta)
require(dplyr)
require(plyr)
require(boot)
#require(reshape2)

### Source plotting function
source("./fig-plot.Occdet_G.R")

# Change ModelNames vetor to species names with spaces
spp_names <- gsub(ModelNames, patt = "_", repl=" ")
spp_names <- sub(spp_names, patt = "c.album", repl="c-album")

# Name the lists of outputs their species names
names(mixLL_list)[1:12] <- spp_names
names(mixLL2_list)[1:12] <- spp_names


# Create a dictionary translating Latin names to common names
dict <- new_lama_dictionary(common_names = 
                              c("Aglais urticae" = 'Small Tortoiseshell',
                                "Carterocephalus palaemon" = 'Chequered Skipper',
                                "Erebia aethiops" = 'Scotch Argus',
                                "Erebia epiphron" = 'Mountain Ringlet',
                                "Limenitis camilla" = 'White Admiral',
                                "Lycaena phlaeas" = 'Small Copper',
                                "Melitaea athalia" = 'Heath Fritillary',
                                "Pararge aegeria" = 'Speckled Wood',
                                "Pieris rapae" = 'Small White',
                                "Polygonia c-album" = 'Comma',
                                "Polyommatus bellargus" = 'Adonis Blue',
                                "Thecla betulae" = 'Brown Hairstreak'))

source("./fig-plot.Occdet_G.R")
```


```{r plot-detprob-outputs-setup, message=FALSE, warning=FALSE, eval=FALSE, echo=FALSE}
require(ggplot2)
require(labelmachine)
require(sparta)
require(dplyr)
require(plyr)
require(boot)


# Change ModelNames vetor to species names with spaces
spp_names <- gsub(ModelNames, patt = "_", repl=" ")
spp_names <- sub(spp_names, patt = "c.album", repl="c-album")

# Name the lists of outputs their species names
names(mixLL_list)[1:12] <- spp_names
names(mixLL2_list)[1:12] <- spp_names


# Create a dictionary translating Latin names to common names
dict <- new_lama_dictionary(common_names = 
                              c("Aglais urticae" = 'Small Tortoiseshell',
                                "Carterocephalus palaemon" = 'Chequered Skipper',
                                "Erebia aethiops" = 'Scotch Argus',
                                "Erebia epiphron" = 'Mountain Ringlet',
                                "Limenitis camilla" = 'White Admiral',
                                "Lycaena phlaeas" = 'Small Copper',
                                "Melitaea athalia" = 'Heath Fritillary',
                                "Pararge aegeria" = 'Speckled Wood',
                                "Pieris rapae" = 'Small White',
                                "Polygonia c-album" = 'Comma',
                                "Polyommatus bellargus" = 'Adonis Blue',
                                "Thecla betulae" = 'Brown Hairstreak'))

```

## *Aglais urticae* (Small Tortoiseshell)

### Occupancy
```{r plot-occu-outputs-A-urticae, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE}
### Source plotting function
source("./fig-plot.Occdet_G.R")

plot.occDet_G(catLL_list[[1]], mixLL_list[[1]], mixLL2_list[[1]]) +
  ggtitle(label = "Aglais urticae", 
          subtitle = paste(lama_translate_("Aglais urticae", dict, "common_names"))) + 
  theme(plot.title=element_text(face="italic")) +
  scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  theme(text=element_text(size=15)) 
```

     
        
        
        
        

*Aglais urticae* (Small Tortoiseshell) is one of the UKs most widespread butterflies, occurring throughout the British Isles, which is in line with the latter part of the time series (post 1976; but note it has declined by approx 15% since). Nonetheless, there is no anecdotal evidence for a 100% increase between the 1960s and 1980s indicating that the occupancy is underestimated in the early part of the time series.   
   
   
   
**In other words it appears neither of the model formulations are accurately explaining early trends for this species**


   

### Detection probability
```{r plot-detprob-outputs-A-urticae, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE, fig.width=7, fig.height=3}
plot_DetectionOverTime(catLL_list[[1]],
                              min.yr = 1900, 
                                    legend_labels =
                         c("    BNM/Year_Effect  ", 
                                                    "UKBMS", 
                                                   "NHCs"),
                                  legend_title = "Data Type") +
             scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
             scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model A") +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank()) +
  #ggtitle(label = (expression(paste(italic("Aglais urticae"), " - Model A")))) #, 
  #        subtitle = paste(lama_translate_("Aglais urticae", dict, "common_names"))) + 
  #theme(plot.title=element_text(face="italic")) +
  theme(text=element_text(size=12)) 

plot_DetectionOverTime(mixLL_list[[1]], 
                       min.yr = 1900, 
                       legend_labels = c("BNM_LL1/Year effect", 
                                         "UKBMS", 
                                         "NHCs", 
                                         "BNM_LL5"), 
                       legend_title = "Data Type") +
  scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model B") +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank()) +
  theme(text=element_text(size=12))

plot_DetectionOverTime(mixLL2_list[[1]],
                              min.yr = 1900, 
                       legend_labels = c("BNM_LL1/Year effect", 
                                         "UKBMS", 
                                         "NHCs", 
                                         "BNM_LL5"), 
                       mixLL2 = TRUE,                  ### Important: specify its mixLL2
                       legend_title = "Data Type") +
  scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model C") +
  theme(text=element_text(size=12))
```


      
      
      
      
BNM and UKBMS detection probability trends are similar for all three models after 1976. However, NHCs detection probability estimates differ between models such that it is the data type with the highest detection probability (unlikely) in Model A, and the lowest in model C; still the detection probability estimates are around 0.07 for model C. If we assume that all NHC visits produced single species lists (which the majority did), this would mean that every 5-10% of museum visits collected a small tortoiseshell, which again seems unlikely as that in turn would mean that approx. the same percentage of the entire collection would be small tortoiseshells).



### PPCs: Visits with records

```{r plot-siteprop-A-urticae, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE, fig.width=7, fig.height=4}
sppMeanSitesYear$Species <- as.factor(sppMeanSitesYear$Species)
test <- subset(sppMeanSitesYear, Species == "*Aglais urticae*")
test[349:464, "Model"] <- "Simulated Data"
test[349:464, "sim_MeanSitesPerYear"] <- test[1:116,"obs_meanSitesPerYear"]
test[349:464, "Year"] <- 1900:2015
test$Model <- as.factor(test$Model)

library(plyr)
test$Model <- revalue(test$Model, c("catLL"="Model A", "mixLL"="Model B", "mixLL2"="Model C"))



ggplot(data = test, aes(x = Year, y = sim_MeanSitesPerYear)) + # Specify data, x and y 
  # Add 95% CIs
  geom_ribbon(aes(fill = Model, # Fill by 'Model'
                  ymin = sim_lower, ymax = sim_upper), # Specify max and min values
                  alpha = 0.3) + # transparency
  scale_fill_manual(values =c("#440154FF", "#20A387FF", # Specify colours
                              "#E69F00", NA)) + # None for last level
  # Add averaged lines
  geom_line(aes(colour = Model), size=0.7) +  # Colour by 'Model'
  scale_colour_manual(values =c("#440154FF", "#20A387FF",  # Specify colours
                                "#E69F00", "black")) +
  # Add shape lines
  geom_point(aes(shape = Model), colour = "black", size = 2) +
  scale_shape_manual(values = c(NA, NA, NA, 15)) + # Only for last level
  # Vertical line at year 1976
  geom_vline(xintercept = 1976, linetype = "dashed") +
  theme_bw() +
    # Tidy legend
  theme(legend.position="bottom", 
        legend.box = "horizontal", 
        legend.title = element_blank()) +
  # Tidy window (no empty space)
  scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
  scale_x_continuous(breaks=seq(1900, 2020, 20), 
                     limits = c(1900, 2022), 
                     expand = c(0, 0)) +
  # Tidy plot appearance
  ylab("Proportion of visits") +
  xlab("Year") +
  ggtitle("Posterior Predictive Checks", subtitle="Small tortoiseshell") + 
  theme(panel.grid = element_blank(),
        text=element_text(size=30), 
        axis.text.x = element_text(angle = 45, hjust = 1))
```

```{r plot-siteprop-A-urticae, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE, fig.width=7, fig.height=4}
sppMeanSitesYear$Species <- as.factor(sppMeanSitesYear$Species)
test <- subset(sppMeanSitesYear, Species == "*Aglais urticae*")

test$Model <- as.factor(test$Model)


require("ggtext")
ggplot(data = test) +
  geom_ribbon(aes( fill = "Model",
                    x = "Year", 
                    y = "sim_MeanSitesPerYear", 
                             ymin = "sim_lower", ymax = "sim_upper"),
                  alpha = 0.3) +
  scale_fill_manual(values =c("#440154FF",
                                    "#20A387FF",
                                    "#E69F00"), 
                          labels = c("A",
                                     "B", 
                                     "C")) + 
  geom_line(aes(x = Year, y = sim_MeanSitesPerYear, colour = Model), size=0.7) + 
          scale_colour_manual(values =c("#440154FF",
                                    "#20A387FF",
                                    "#E69F00"), 
                          labels = c("A",
                                     "B", 
                                     "C")) +
  geom_line(aes(x=Year, y=obs_meanSitesPerYear), linetype = "dotted", colour=c("black"), size = 1.5) +
  geom_line(aes(x=Year, y=obs_meanSitesPerYear), colour=c("black"), size = 0.4) +
          #scale_colour_manual(name = 'Simulated Data', 
          #                    values =c("black"), 
          #                labels = c("")) +
  geom_vline(xintercept = 1976, linetype = "dashed") +   # Vertical line = 1976
  ggtitle("Aglais urticae") + 
  #theme(plot.title=element_text(face="italic")) +
  ylab("Proportion of visited sites with records") +
  xlab("Year") +
  theme_bw() +
  #guides(Model = guide_legend(order = 1),
  #       obs_meanSitesPerYear = guide_legend(order = 2)) +
  scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
  scale_x_continuous(breaks=seq(1900, 2020, 20), 
                     limits = c(1900, 2025), 
                     expand = c(0, 0)) +
      #theme(legend.position="bottom", legend.box = "horizontal") + 
  labs(subtitle="Legend at Bottom") 
```


### Record numbers
```{r plot-record-numbers-A-urticae, message=FALSE, warning=FALSE, echo=FALSE, fig.width=7, fig.height=4}
# Load required packages
require(dplyr)
require(plyr)
require(ggplot2)
require(viridis)
require(readr)
require(ggtext)

# Load spp_vis data (any would do), which contain all visits and whether or not each of the three species have been observed at that visit
spp_vis <- read.csv('../data/formattedData/spp_vis_merged_2020-04-07.csv', header=T, na.strings=c("","NA"))

# Load occDetdata giving each visits' site ID (grid cell), list length (numper of species observed) and time period (in this case year)
occDetdata_merged_2020_04_07 <- read.csv('../data/formattedData/occDetdata_merged_2020-04-07.csv', header=T, na.strings=c("","NA"))

# Merge by visit
spp_vis <- merge(spp_vis, occDetdata_merged_2020_04_07, by="visit")

# Find unique visits
spp_vis <- unique(spp_vis)

# Create a species column
spp_vis$species <- as.factor(c("NA"))

### Make L column a source column
# Change all List lengths between 0 and 22222 to 1 (all these are BNM data)
spp_vis$L <- replace(spp_vis$L, spp_vis$L > 0 & spp_vis$L < 22222, 1)
# Make L a factor
spp_vis$L <- as.factor(spp_vis$L)
# Rename factor levels according to data type
spp_vis$L <- revalue(spp_vis$L, c("1"="BNM", "22222"="UKBMS", "33333"="NHCs"))
# Rename column to Source
colnames(spp_vis)[colnames(spp_vis)=="L"] <- "Source"


# Correct time period so it starts at 1900
spp_vis$year <- spp_vis$TP+1899



########################################################################

# Subset each species' records into dataframes and specify species identity
spp_vis_A_urticae <- subset(spp_vis, Aglais.urticae == "TRUE")
spp_vis_A_urticae$species <- c("italic('A. urticae')")



 ggplot(spp_vis_A_urticae, aes(x=year, fill=Source, color=Source)) + 
  scale_fill_manual(values = alpha(c("#453781FF", 
                                     "#20A386FF", 
                                     "#DCE318FF"), 0.5)) +
  scale_color_manual(values = alpha(c("#453781FF", 
                                      "#20A386FF", 
                                      "#DCE318FF"), 0.5)) +
  geom_bar(position="identity", alpha=0.7) + 
  labs(y="Records", x = "Year") +
  theme_bw() +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank()) +
  # sqrt y scale and expand scale to no space below, 5% above
  scale_y_sqrt(expand = expansion(mult = c(0, .05))) +
  scale_x_continuous(breaks=seq(1900, 2020, 20), 
                     limits = c(1900, 2025), 
                     expand = c(0, 0)) +
  ggtitle("Aglais urticae") + 
  theme(plot.title=element_text(face="italic")) +
  theme(text = element_text(size=15),
        axis.text.x = element_text(angle=90, hjust=1))


```



### DIC & Bayesian *p*
```{r DIC-A-urticae, cache=TRUE, message=FALSE, echo=FALSE}
# Create empty data frame to populate
summary_out <- data.frame(Model = as.factor(c("catLL", "mixLL", "mixLL2")),
                          DIC = as.numeric(NA),
                          Bayesian_p = as.numeric(NA),
                          Bayesian_p_var = as.numeric(NA)) 


  summary_out[1:3, "DIC"] <-  c((catLL_list[[1]]$BUGSoutput$DIC), 
                                (mixLL_list[[1]]$BUGSoutput$DIC),
                                (mixLL2_list[[1]]$BUGSoutput$DIC))  
  summary_out[1, "Bayesian_p"] <-  ModelBp_Summ[1, 2]
  summary_out[2, "Bayesian_p"] <-  ModelBp_Summ[1, 3]
  summary_out[3, "Bayesian_p"] <-  ModelBp_Summ[1, 4]
  summary_out[1, "Bayesian_p_var"] <-  ModelBp_Summ[1, 5]
  summary_out[2, "Bayesian_p_var"] <-  ModelBp_Summ[1, 6]
  summary_out[3, "Bayesian_p_var"] <-  ModelBp_Summ[1, 7]

knitr::kable(summary_out)
```






* Bayesian p:     
    + very poor for variance      
    + OK for mean    
    + Model A (Cat list length) best    
* DIC: Model C (mic list length 2) best   







## *Carterocephalus palaemon* (Chequered skipper)

### Occupancy
```{r plot-occu-outputs-C-palaemon, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE}
### Source plotting function
source("./fig-plot.Occdet_G.R")

plot.occDet_G(mixLL_list[[2]], mixLL2_list[[2]], catLL_list[[2]]) +
  ggtitle(label = "Carterocephalus palaemon", 
          subtitle = paste(lama_translate_("Carterocephalus palaemon", dict, "common_names"))) + 
  theme(plot.title=element_text(face="italic")) +
  scale_y_continuous(limits=c(0, 0.1), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  theme(text=element_text(size=15))  
```
   
   
   
   
The Chequered Skipper is confined to north-west Scotland where it was first discovered in 1939 (distribution is centered on Fort William). It formerly occurred in central and east England, but started to decline in "the early  part of the 20th century" (Millennium atlas) with East Midlands as the last English stronghold where it was "fairly abundant until the 1950s". Thereafter remaining English populations suffered rapid declines and it was declared extinct in England in 1976. 
   
   
**Occupancy estimates are in line with anecdotal evidence for historic distribution changes and very similar across models**   
    
   
### Detection probability
```{r plot-detprob-outputs-C-palaemon, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE, fig.width=7, fig.height=3}
plot_DetectionOverTime(catLL_list[[2]],
                              min.yr = 1900, 
                                    legend_labels =
                         c("    BNM/Year_Effect  ", 
                                                    "UKBMS", 
                                                   "NHCs"),
                                  legend_title = "Data Type") +
             scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
             scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model A") +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank()) +
  theme(text=element_text(size=12)) 

plot_DetectionOverTime(mixLL_list[[2]], 
                       min.yr = 1900, 
                       legend_labels = c("BNM_LL1/Year effect", 
                                         "UKBMS", 
                                         "NHCs", 
                                         "BNM_LL5"), 
                       legend_title = "Data Type") +
  scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model B") +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank()) +
  theme(text=element_text(size=12))

plot_DetectionOverTime(mixLL2_list[[2]],
                              min.yr = 1900, 
                       legend_labels = c("BNM_LL1/Year effect", 
                                         "UKBMS", 
                                         "NHCs", 
                                         "BNM_LL5"), 
                       mixLL2 = TRUE,                  ### Important: specify its mixLL2
                       legend_title = "Data Type") +
  scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model C") +
  theme(text=element_text(size=12))

```
   
   
   
   
Detection probabilities when compared between models are similar to previous species. However, I would expect detection probabilities to be higher? Although it is quite hard to spot (skipper = rapid flier etc.). BUT I would expect the detection probability for NHCs in Model C to be much higher than it is.... 
   
   
**General thought: detection probability does not appear to influence occupancy estimates that much?**   
    
   



### Visits with records
```{r plot-siteprop-C-palaemon, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE, fig.width=7, fig.height=4}
sppMeanSitesYear$Species <- as.factor(sppMeanSitesYear$Species)
test <- subset(sppMeanSitesYear, Species == "*Carterocephalus palaemon*")

test$Model <- as.factor(test$Model)


require("ggtext")
ggplot(data = test,) +
  geom_ribbon(aes_string(fill = "Model", 
                    x = "Year", 
                    y = "sim_MeanSitesPerYear", 
                             ymin = "sim_lower", ymax = "sim_upper"),
                  alpha = 0.3) +
  scale_fill_manual(values =c("#440154FF", "#20A387FF", "#E69F00")) + 
  geom_line(aes(x = Year, y = sim_MeanSitesPerYear, colour = Model), size=0.7) + 
          scale_colour_manual(values =c("#440154FF", "#20A387FF", "#E69F00")) +
  geom_line(aes(x=Year, y=obs_meanSitesPerYear), linetype = "dotted", colour=c("black"), size = 1.5) +
  geom_line(aes(x=Year, y=obs_meanSitesPerYear), colour=c("black"), size = 0.4) +
  geom_vline(xintercept = 1976, linetype = "dashed") +   # Vertical line = 1976
  ylab("Proportion of visited sites with records") +
  xlab("Year") +
  scale_y_continuous(limits=c(0, 0.1), expand = c(0, 0)) +
  scale_x_continuous(breaks=seq(1900, 2020, 20), 
                     limits = c(1900, 2025), 
                     expand = c(0, 0)) +
      #theme(legend.position="bottom") +
  xlab("Year") + 
  ylab("Proportion of visited sites with records") + 
  ggtitle("Carterocephalus palaemon") + 
  theme(plot.title=element_text(face="italic")) +
  theme_bw()

```


### Record numbers
```{r plot-record-numbers-C-palaemon, message=FALSE, warning=FALSE, echo=FALSE, fig.width=7, fig.height=4, cache=TRUE}
# Subset each species' records into dataframes and specify species identity
spp_vis_C_palaemon <- subset(spp_vis, Carterocephalus.palaemon == "TRUE")
spp_vis_C_palaemon$species <- c("italic('C. palaemon')")

 ggplot(spp_vis_C_palaemon, aes(x=year, fill=Source, color=Source)) + 
  scale_fill_manual(values = alpha(c("#453781FF", 
                                     "#20A386FF", 
                                     "#DCE318FF"), 0.5)) +
  scale_color_manual(values = alpha(c("#453781FF", 
                                      "#20A386FF", 
                                      "#DCE318FF"), 0.5)) +
  geom_bar(position="identity", alpha=0.7) + 
  labs(y="Records", x = "Year") +
  theme_bw() +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank()) +
  # sqrt y scale and expand scale to no space below, 5% above
  scale_y_sqrt(expand = expansion(mult = c(0, .05))) +
  scale_x_continuous(breaks=seq(1900, 2020, 20), 
                     limits = c(1900, 2025), 
                     expand = c(0, 0)) +
  ggtitle("Carterocephalus palaemon") + 
  theme(plot.title=element_text(face="italic")) +
  theme(text = element_text(size=15),
        axis.text.x = element_text(angle=90, hjust=1))
 
```


### DIC & Bayesian *p*
```{r DIC-C-palaemon, cache=TRUE, message=FALSE, echo=FALSE, cache=TRUE}
# Create empty data frame to populate
summary_out <- data.frame(Model = as.factor(c("catLL", "mixLL", "mixLL2")),
                          DIC = as.numeric(NA),
                          Bayesian_p = as.numeric(NA),
                          Bayesian_p_var = as.numeric(NA)) 


  summary_out[1:3, "DIC"] <-  c((catLL_list[[2]]$BUGSoutput$DIC), 
                                (mixLL_list[[2]]$BUGSoutput$DIC),
                                (mixLL2_list[[2]]$BUGSoutput$DIC))  
  summary_out[1, "Bayesian_p"] <-  ModelBp_Summ[2, 2]
  summary_out[2, "Bayesian_p"] <-  ModelBp_Summ[2, 3]
  summary_out[3, "Bayesian_p"] <-  ModelBp_Summ[2, 4]
  summary_out[1, "Bayesian_p_var"] <-  ModelBp_Summ[2, 5]
  summary_out[2, "Bayesian_p_var"] <-  ModelBp_Summ[2, 6]
  summary_out[3, "Bayesian_p_var"] <-  ModelBp_Summ[2, 7]

knitr::kable(summary_out)
```


* Bayesian p:     
    + Poor for variance: A best      
    + OK for mean: C best          
* DIC: Model B (mic list length) best   




## *Erebia aethiops* (Scotch Argus)

### Occupancy
```{r plot-occu-outputs-E-aethiops, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE}
### Source plotting function
source("./fig-plot.Occdet_G.R")

plot.occDet_G(mixLL_list[[3]], mixLL2_list[[3]], catLL_list[[3]]) +
  ggtitle(label = "Erebia aethiops", 
          subtitle = paste(lama_translate_("Erebia aethiops", dict, "common_names"))) + 
  theme(plot.title=element_text(face="italic")) +
  scale_y_continuous(limits=c(0, 0.25), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  theme(text=element_text(size=15))  
```


### Detection probability
```{r plot-detprob-outputs-E-aethiops, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE, fig.width=7, fig.height=3}
plot_DetectionOverTime(catLL_list[[3]],
                              min.yr = 1900, 
                                    legend_labels =
                         c("    BNM/Year_Effect  ", 
                                                    "UKBMS", 
                                                   "NHCs"),
                                  legend_title = "Data Type") +
             scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
             scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model A") +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank()) +
  theme(text=element_text(size=12)) 

plot_DetectionOverTime(mixLL_list[[3]], 
                       min.yr = 1900, 
                       legend_labels = c("BNM_LL1/Year effect", 
                                         "UKBMS", 
                                         "NHCs", 
                                         "BNM_LL5"), 
                       legend_title = "Data Type") +
  scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model B") +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank()) +
  theme(text=element_text(size=12))

plot_DetectionOverTime(mixLL2_list[[3]],
                              min.yr = 1900, 
                       legend_labels = c("BNM_LL1/Year effect", 
                                         "UKBMS", 
                                         "NHCs", 
                                         "BNM_LL5"), 
                       mixLL2 = TRUE,                  ### Important: specify its mixLL2
                       legend_title = "Data Type") +
  scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model C") +
  theme(text=element_text(size=12))

```



### Visits with records
```{r plot-siteprop-E-aethiops, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE, fig.width=7, fig.height=4}
sppMeanSitesYear$Species <- as.factor(sppMeanSitesYear$Species)
test <- subset(sppMeanSitesYear, Species == "*Erebia aethiops*")

test$Model <- as.factor(test$Model)

require("ggtext")
ggplot(data = test,) +
  geom_ribbon(aes_string(fill = "Model", 
                    x = "Year", 
                    y = "sim_MeanSitesPerYear", 
                             ymin = "sim_lower", ymax = "sim_upper"), alpha = 0.3) +
  scale_fill_manual(values =c("#440154FF", "#20A387FF", "#E69F00")) + 
  geom_line(aes(x = Year, y = sim_MeanSitesPerYear, colour = Model), size=0.7) + 
          scale_colour_manual(values =c("#440154FF", "#20A387FF", "#E69F00")) +
  geom_line(aes(x=Year, y=obs_meanSitesPerYear), linetype = "dotted", colour=c("black"), size = 1.5) +
  geom_line(aes(x=Year, y=obs_meanSitesPerYear), colour=c("black"), size = 0.4) +
  geom_vline(xintercept = 1976, linetype = "dashed") +   # Vertical line = 1976
  ylab("Proportion of visited sites with records") + xlab("Year") +
  scale_y_continuous(limits=c(0, 0.1), expand = c(0, 0)) +
  scale_x_continuous(breaks=seq(1900, 2020, 20), 
                     limits = c(1900, 2025), 
                     expand = c(0, 0)) +
  ggtitle("Erebia aethiops") + 
  theme(plot.title=element_text(face="italic")) +
  theme_bw()
```


### Record numbers
```{r plot-record-numbers-E-aethiops, message=FALSE, warning=FALSE, echo=FALSE, fig.width=7, fig.height=4, cache=TRUE}
# Subset each species' records into dataframes and specify species identity
spp_vis_E_aethiops <- subset(spp_vis, Erebia.aethiops == "TRUE")
spp_vis_E_aethiops$species <- c("italic('E. aethiops')")

 ggplot(spp_vis_E_aethiops, aes(x=year, fill=Source, color=Source)) + 
  scale_fill_manual(values = alpha(c("#453781FF", 
                                     "#20A386FF", 
                                     "#DCE318FF"), 0.5)) +
  scale_color_manual(values = alpha(c("#453781FF", 
                                      "#20A386FF", 
                                      "#DCE318FF"), 0.5)) +
  geom_bar(position="identity", alpha=0.7) + 
  labs(y="Records", x = "Year") +
  theme_bw() +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank()) +
  # sqrt y scale and expand scale to no space below, 5% above
  scale_y_sqrt(expand = expansion(mult = c(0, .05))) +
  scale_x_continuous(breaks=seq(1900, 2020, 20), 
                     limits = c(1900, 2025), 
                     expand = c(0, 0)) +
  ggtitle("Erebia aethiops") + 
  theme(plot.title=element_text(face="italic")) +
  theme(text = element_text(size=15),
        axis.text.x = element_text(angle=90, hjust=1))
```


### DIC & Bayesian *p*
```{r DIC-E-aethiops, cache=TRUE, message=FALSE, echo=FALSE, cache=TRUE}
# Create empty data frame to populate
summary_out <- data.frame(Model = as.factor(c("catLL", "mixLL", "mixLL2")),
                          DIC = as.numeric(NA),
                          Bayesian_p = as.numeric(NA),
                          Bayesian_p_var = as.numeric(NA)) 


  summary_out[1:3, "DIC"] <-  c((catLL_list[[3]]$BUGSoutput$DIC), 
                                (mixLL_list[[3]]$BUGSoutput$DIC),
                                (mixLL2_list[[3]]$BUGSoutput$DIC))  
  summary_out[1, "Bayesian_p"] <-  ModelBp_Summ[3, 2]
  summary_out[2, "Bayesian_p"] <-  ModelBp_Summ[3, 3]
  summary_out[3, "Bayesian_p"] <-  ModelBp_Summ[3, 4]
  summary_out[1, "Bayesian_p_var"] <-  ModelBp_Summ[3, 5]
  summary_out[2, "Bayesian_p_var"] <-  ModelBp_Summ[3, 6]
  summary_out[3, "Bayesian_p_var"] <-  ModelBp_Summ[3, 7]

knitr::kable(summary_out)
```




## *Erebia epiphron* (Mountain Ringlet)

### Occupancy
```{r plot-occu-outputs-E-epiphron, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE}
### Source plotting function
source("./fig-plot.Occdet_G.R")

plot.occDet_G(mixLL_list[[4]], mixLL2_list[[4]], catLL_list[[4]]) +
  ggtitle(label = "Erebia epiphron", 
          subtitle = paste(lama_translate_("Erebia epiphron", dict, "common_names"))) + 
  theme(plot.title=element_text(face="italic")) +
  scale_y_continuous(limits=c(0, 0.1), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  theme(text=element_text(size=15)) 
```



### Detection probability
```{r plot-detprob-outputs-E-epiphron, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE, fig.width=7, fig.height=3}
plot_DetectionOverTime(catLL_list[[4]],
                              min.yr = 1900, 
                                    legend_labels =
                         c("    BNM/Year_Effect  ", 
                                                    "UKBMS", 
                                                   "NHCs"),
                                  legend_title = "Data Type") +
             scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
             scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model A") +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank()) +
  theme(text=element_text(size=12)) 

plot_DetectionOverTime(mixLL_list[[4]], 
                       min.yr = 1900, 
                       legend_labels = c("BNM_LL1/Year effect", 
                                         "UKBMS", 
                                         "NHCs", 
                                         "BNM_LL5"), 
                       legend_title = "Data Type") +
  scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model B") +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank()) +
  theme(text=element_text(size=12))

plot_DetectionOverTime(mixLL2_list[[4]],
                              min.yr = 1900, 
                       legend_labels = c("BNM_LL1/Year effect", 
                                         "UKBMS", 
                                         "NHCs", 
                                         "BNM_LL5"), 
                       mixLL2 = TRUE,                  ### Important: specify its mixLL2
                       legend_title = "Data Type") +
  scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model C") +
  theme(text=element_text(size=12))

```



### Visits with records
```{r plot-siteprop-E-epiphron, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE, fig.width=7, fig.height=4}
sppMeanSitesYear$Species <- as.factor(sppMeanSitesYear$Species)
test <- subset(sppMeanSitesYear, Species == "*Erebia epiphron*")

test$Model <- as.factor(test$Model)

require("ggtext")
ggplot(data = test,) +
  geom_ribbon(aes_string(fill = "Model", 
                    x = "Year", 
                    y = "sim_MeanSitesPerYear", 
                             ymin = "sim_lower", ymax = "sim_upper"), alpha = 0.3) +
  scale_fill_manual(values =c("#440154FF", "#20A387FF", "#E69F00")) + 
  geom_line(aes(x = Year, y = sim_MeanSitesPerYear, colour = Model), size=0.7) + 
          scale_colour_manual(values =c("#440154FF", "#20A387FF", "#E69F00")) +
  geom_line(aes(x=Year, y=obs_meanSitesPerYear), linetype = "dotted", colour=c("black"), size = 1.5) +
  geom_line(aes(x=Year, y=obs_meanSitesPerYear), colour=c("black"), size = 0.4) +
  geom_vline(xintercept = 1976, linetype = "dashed") +   # Vertical line = 1976
  ylab("Proportion of visited sites with records") + xlab("Year") +
  scale_y_continuous(limits=c(0, 0.1), expand = c(0, 0)) +
  scale_x_continuous(breaks=seq(1900, 2020, 20), 
                     limits = c(1900, 2025), 
                     expand = c(0, 0)) +
  ggtitle("Erebia epiphron") + 
  theme(plot.title=element_text(face="italic")) +
  theme_bw()
```


### Record numbers
```{r plot-record-numbers-E-epiphron, message=FALSE, warning=FALSE, echo=FALSE, fig.width=7, fig.height=4, cache=TRUE}
# Subset each species' records into dataframes and specify species identity
spp_vis_E_epiphron <- subset(spp_vis, Erebia.epiphron == "TRUE")
spp_vis_E_epiphron$species <- c("italic('E. epiphron')")

 ggplot(spp_vis_E_epiphron, aes(x=year, fill=Source, color=Source)) + 
  scale_fill_manual(values = alpha(c("#453781FF", 
                                     "#20A386FF", 
                                     "#DCE318FF"), 0.5)) +
  scale_color_manual(values = alpha(c("#453781FF", 
                                      "#20A386FF", 
                                      "#DCE318FF"), 0.5)) +
  geom_bar(position="identity", alpha=0.7) + 
  labs(y="Records", x = "Year") +
  theme_bw() +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank()) +
  # sqrt y scale and expand scale to no space below, 5% above
  scale_y_sqrt(expand = expansion(mult = c(0, .05))) +
  scale_x_continuous(breaks=seq(1900, 2020, 20), 
                     limits = c(1900, 2025), 
                     expand = c(0, 0)) +
  ggtitle("Erebia epiphron") + 
  theme(plot.title=element_text(face="italic")) +
  theme(text = element_text(size=15),
        axis.text.x = element_text(angle=90, hjust=1))
```


### DIC & Bayesian *p*
```{r DIC-E-epiphron, cache=TRUE, message=FALSE, echo=FALSE, cache=TRUE}
# Create empty data frame to populate
summary_out <- data.frame(Model = as.factor(c("catLL", "mixLL", "mixLL2")),
                          DIC = as.numeric(NA),
                          Bayesian_p = as.numeric(NA),
                          Bayesian_p_var = as.numeric(NA)) 


  summary_out[1:3, "DIC"] <-  c((catLL_list[[4]]$BUGSoutput$DIC), 
                                (mixLL_list[[4]]$BUGSoutput$DIC),
                                (mixLL2_list[[4]]$BUGSoutput$DIC))  
  summary_out[1, "Bayesian_p"] <-  ModelBp_Summ[4, 2]
  summary_out[2, "Bayesian_p"] <-  ModelBp_Summ[4, 3]
  summary_out[3, "Bayesian_p"] <-  ModelBp_Summ[4, 4]
  summary_out[1, "Bayesian_p_var"] <-  ModelBp_Summ[4, 5]
  summary_out[2, "Bayesian_p_var"] <-  ModelBp_Summ[4, 6]
  summary_out[3, "Bayesian_p_var"] <-  ModelBp_Summ[4, 7]

knitr::kable(summary_out)
```















## *Limenitis camilla* (White Admiral)

### Occupancy
```{r plot-occu-outputs-L-camilla, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE}
### Source plotting function
source("./fig-plot.Occdet_G.R")

plot.occDet_G(mixLL_list[[5]], mixLL2_list[[5]], catLL_list[[5]]) +
  ggtitle(label = "Limenitis camilla", 
          subtitle = paste(lama_translate_("Limenitis camilla", dict, "common_names"))) + 
  theme(plot.title=element_text(face="italic")) +
  scale_y_continuous(limits=c(0, 0.4), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  theme(text=element_text(size=15)) 
```



### Detection probability
```{r plot-detprob-outputs-L-camilla, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE, fig.width=7, fig.height=3}
plot_DetectionOverTime(catLL_list[[5]],
                              min.yr = 1900, 
                                    legend_labels =
                         c("    BNM/Year_Effect  ", 
                                                    "UKBMS", 
                                                   "NHCs"),
                                  legend_title = "Data Type") +
             scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
             scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model A") +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank()) +
  theme(text=element_text(size=12)) 

plot_DetectionOverTime(mixLL_list[[5]], 
                       min.yr = 1900, 
                       legend_labels = c("BNM_LL1/Year effect", 
                                         "UKBMS", 
                                         "NHCs", 
                                         "BNM_LL5"), 
                       legend_title = "Data Type") +
  scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model B") +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank()) +
  theme(text=element_text(size=12))

plot_DetectionOverTime(mixLL2_list[[5]],
                              min.yr = 1900, 
                       legend_labels = c("BNM_LL1/Year effect", 
                                         "UKBMS", 
                                         "NHCs", 
                                         "BNM_LL5"), 
                       mixLL2 = TRUE,                  ### Important: specify its mixLL2
                       legend_title = "Data Type") +
  scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model C") +
  theme(text=element_text(size=12))
```



### Visits with records
```{r plot-siteprop-L-camilla, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE, fig.width=7, fig.height=4}
sppMeanSitesYear$Species <- as.factor(sppMeanSitesYear$Species)
test <- subset(sppMeanSitesYear, Species == "*Limenitis camilla*")

test$Model <- as.factor(test$Model)

require("ggtext")
ggplot(data = test,) +
  geom_ribbon(aes_string(fill = "Model", 
                    x = "Year", 
                    y = "sim_MeanSitesPerYear", 
                             ymin = "sim_lower", ymax = "sim_upper"), alpha = 0.3) +
  scale_fill_manual(values =c("#440154FF", "#20A387FF", "#E69F00")) + 
  geom_line(aes(x = Year, y = sim_MeanSitesPerYear, colour = Model), size=0.7) + 
          scale_colour_manual(values =c("#440154FF", "#20A387FF", "#E69F00")) +
  geom_line(aes(x=Year, y=obs_meanSitesPerYear), linetype = "dotted", colour=c("black"), size = 1.5) +
  geom_line(aes(x=Year, y=obs_meanSitesPerYear), colour=c("black"), size = 0.4) +
  geom_vline(xintercept = 1976, linetype = "dashed") +   # Vertical line = 1976
  ylab("Proportion of visited sites with records") + xlab("Year") +
  scale_y_continuous(limits=c(0, 0.25), expand = c(0, 0)) +
  scale_x_continuous(breaks=seq(1900, 2020, 20), 
                     limits = c(1900, 2025), 
                     expand = c(0, 0)) +
  ggtitle("Limenitis camilla") + 
  theme(plot.title=element_text(face="italic")) +
  theme_bw()
```


### Record numbers
```{r plot-record-numbers-L-camilla, message=FALSE, warning=FALSE, echo=FALSE, fig.width=7, fig.height=4, cache=TRUE}
# Subset each species' records into dataframes and specify species identity
spp_vis_L_camilla <- subset(spp_vis, Limenitis.camilla == "TRUE")
spp_vis_L_camilla$species <- c("italic('L. camilla')")

 ggplot(spp_vis_L_camilla, aes(x=year, fill=Source, color=Source)) + 
  scale_fill_manual(values = alpha(c("#453781FF", 
                                     "#20A386FF", 
                                     "#DCE318FF"), 0.5)) +
  scale_color_manual(values = alpha(c("#453781FF", 
                                      "#20A386FF", 
                                      "#DCE318FF"), 0.5)) +
  geom_bar(position="identity", alpha=0.7) + 
  labs(y="Records", x = "Year") +
  theme_bw() +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank()) +
  # sqrt y scale and expand scale to no space below, 5% above
  scale_y_sqrt(expand = expansion(mult = c(0, .05))) +
  scale_x_continuous(breaks=seq(1900, 2020, 20), 
                     limits = c(1900, 2025), 
                     expand = c(0, 0)) +
  ggtitle("Limenitis camilla") + 
  theme(plot.title=element_text(face="italic")) +
  theme(text = element_text(size=15),
        axis.text.x = element_text(angle=90, hjust=1))
```


### DIC & Bayesian *p*
```{r DIC-L-camilla, cache=TRUE, message=FALSE, echo=FALSE, cache=TRUE}
# Create empty data frame to populate
summary_out <- data.frame(Model = as.factor(c("catLL", "mixLL", "mixLL2")),
                          DIC = as.numeric(NA),
                          Bayesian_p = as.numeric(NA),
                          Bayesian_p_var = as.numeric(NA)) 


  summary_out[1:3, "DIC"] <-  c((catLL_list[[5]]$BUGSoutput$DIC), 
                                (mixLL_list[[5]]$BUGSoutput$DIC),
                                (mixLL2_list[[5]]$BUGSoutput$DIC))  
  summary_out[1, "Bayesian_p"] <-  ModelBp_Summ[5, 2]
  summary_out[2, "Bayesian_p"] <-  ModelBp_Summ[5, 3]
  summary_out[3, "Bayesian_p"] <-  ModelBp_Summ[5, 4]
  summary_out[1, "Bayesian_p_var"] <-  ModelBp_Summ[5, 5]
  summary_out[2, "Bayesian_p_var"] <-  ModelBp_Summ[5, 6]
  summary_out[3, "Bayesian_p_var"] <-  ModelBp_Summ[5, 7]

knitr::kable(summary_out)
```















## *Lycaena phlaeas* (Small copper)

### Occupancy
```{r plot-occu-outputs-L-phlaeas, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE}
### Source plotting function
source("./fig-plot.Occdet_G.R")

plot.occDet_G(mixLL_list[[6]], mixLL2_list[[6]], catLL_list[[6]]) +
  ggtitle(label = "Lycaena phlaeas", 
          subtitle = paste(lama_translate_("Lycaena phlaeas", dict, "common_names"))) + 
  theme(plot.title=element_text(face="italic")) +
  scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  theme(text=element_text(size=15)) 
```



### Detection probability
```{r plot-detprob-outputs-L-phlaeas, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE, fig.width=7, fig.height=3}
plot_DetectionOverTime(catLL_list[[6]],
                              min.yr = 1900, 
                                    legend_labels =
                         c("    BNM/Year_Effect  ", 
                                                    "UKBMS", 
                                                   "NHCs"),
                                  legend_title = "Data Type") +
             scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
             scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model A") +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank()) +
  theme(text=element_text(size=12)) 

plot_DetectionOverTime(mixLL_list[[6]], 
                       min.yr = 1900, 
                       legend_labels = c("BNM_LL1/Year effect", 
                                         "UKBMS", 
                                         "NHCs", 
                                         "BNM_LL5"), 
                       legend_title = "Data Type") +
  scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model B") +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank()) +
  theme(text=element_text(size=12))

plot_DetectionOverTime(mixLL2_list[[6]],
                              min.yr = 1900, 
                       legend_labels = c("BNM_LL1/Year effect", 
                                         "UKBMS", 
                                         "NHCs", 
                                         "BNM_LL5"), 
                       mixLL2 = TRUE,                  ### Important: specify its mixLL2
                       legend_title = "Data Type") +
  scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model C") +
  theme(text=element_text(size=12))
```



### Visits with records
```{r plot-siteprop-L-phlaeas, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE, fig.width=7, fig.height=4}
sppMeanSitesYear$Species <- as.factor(sppMeanSitesYear$Species)
test <- subset(sppMeanSitesYear, Species == "*Lycaena phlaeas*")

test$Model <- as.factor(test$Model)

require("ggtext")
ggplot(data = test,) +
  geom_ribbon(aes_string(fill = "Model", 
                    x = "Year", 
                    y = "sim_MeanSitesPerYear", 
                             ymin = "sim_lower", ymax = "sim_upper"), alpha = 0.3) +
  scale_fill_manual(values =c("#440154FF", "#20A387FF", "#E69F00")) + 
  geom_line(aes(x = Year, y = sim_MeanSitesPerYear, colour = Model), size=0.7) + 
          scale_colour_manual(values =c("#440154FF", "#20A387FF", "#E69F00")) +
  geom_line(aes(x=Year, y=obs_meanSitesPerYear), linetype = "dotted", colour=c("black"), size = 1.5) +
  geom_line(aes(x=Year, y=obs_meanSitesPerYear), colour=c("black"), size = 0.4) +
  geom_vline(xintercept = 1976, linetype = "dashed") +   # Vertical line = 1976
  ylab("Proportion of visited sites with records") + xlab("Year") +
  scale_y_continuous(limits=c(0, 0.8), expand = c(0, 0)) +
  scale_x_continuous(breaks=seq(1900, 2020, 20), 
                     limits = c(1900, 2025), 
                     expand = c(0, 0)) +
  ggtitle("Lycaena phlaeas") + 
  theme(plot.title=element_text(face="italic")) +
  theme_bw()
```


### Record numbers
```{r plot-record-numbers-L-phlaeas, message=FALSE, warning=FALSE, echo=FALSE, fig.width=7, fig.height=4, cache=TRUE}
# Subset each species' records into dataframes and specify species identity
spp_vis_L_phlaeas <- subset(spp_vis, Lycaena.phlaeas == "TRUE")
spp_vis_L_phlaeas$species <- c("italic('L. phlaeas')")

 ggplot(spp_vis_L_phlaeas, aes(x=year, fill=Source, color=Source)) + 
  scale_fill_manual(values = alpha(c("#453781FF", 
                                     "#20A386FF", 
                                     "#DCE318FF"), 0.5)) +
  scale_color_manual(values = alpha(c("#453781FF", 
                                      "#20A386FF", 
                                      "#DCE318FF"), 0.5)) +
  geom_bar(position="identity", alpha=0.7) + 
  labs(y="Records", x = "Year") +
  theme_bw() +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank()) +
  # sqrt y scale and expand scale to no space below, 5% above
  scale_y_sqrt(expand = expansion(mult = c(0, .05))) +
  scale_x_continuous(breaks=seq(1900, 2020, 20), 
                     limits = c(1900, 2025), 
                     expand = c(0, 0)) +
  ggtitle("Lycaena phlaeas") + 
  theme(plot.title=element_text(face="italic")) +
  theme(text = element_text(size=15),
        axis.text.x = element_text(angle=90, hjust=1))
```


### DIC & Bayesian *p*
```{r DIC-L-phlaeas, cache=TRUE, message=FALSE, echo=FALSE, cache=TRUE}
# Create empty data frame to populate
summary_out <- data.frame(Model = as.factor(c("catLL", "mixLL", "mixLL2")),
                          DIC = as.numeric(NA),
                          Bayesian_p = as.numeric(NA),
                          Bayesian_p_var = as.numeric(NA)) 


  summary_out[1:3, "DIC"] <-  c((catLL_list[[6]]$BUGSoutput$DIC), 
                                (mixLL_list[[6]]$BUGSoutput$DIC),
                                (mixLL2_list[[6]]$BUGSoutput$DIC))  
  summary_out[1, "Bayesian_p"] <-  ModelBp_Summ[6, 2]
  summary_out[2, "Bayesian_p"] <-  ModelBp_Summ[6, 3]
  summary_out[3, "Bayesian_p"] <-  ModelBp_Summ[6, 4]
  summary_out[1, "Bayesian_p_var"] <-  ModelBp_Summ[6, 5]
  summary_out[2, "Bayesian_p_var"] <-  ModelBp_Summ[6, 6]
  summary_out[3, "Bayesian_p_var"] <-  ModelBp_Summ[6, 7]

knitr::kable(summary_out)
```















## *Melitaea athalia* (Heath Fritillary)

### Occupancy
```{r plot-occu-outputs-M-athalia, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE}
### Source plotting function
source("./fig-plot.Occdet_G.R")

plot.occDet_G(mixLL_list[[7]], mixLL2_list[[7]], catLL_list[[7]]) +
  ggtitle(label = "Melitaea athalia", 
          subtitle = paste(lama_translate_("Melitaea athalia", dict, "common_names"))) + 
  theme(plot.title=element_text(face="italic")) +
  scale_y_continuous(limits=c(0, 0.1), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  theme(text=element_text(size=15)) 
```



### Detection probability
```{r plot-detprob-outputs-M-athalia, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE, fig.width=7, fig.height=3}

plot_DetectionOverTime(catLL_list[[7]],
                              min.yr = 1900, 
                                    legend_labels =
                         c("    BNM/Year_Effect  ", 
                                                    "UKBMS", 
                                                   "NHCs"),
                                  legend_title = "Data Type") +
             scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
             scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model A") +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank()) +
  theme(text=element_text(size=12)) 

plot_DetectionOverTime(mixLL_list[[7]], 
                       min.yr = 1900, 
                       legend_labels = c("BNM_LL1/Year effect", 
                                         "UKBMS", 
                                         "NHCs", 
                                         "BNM_LL5"), 
                       legend_title = "Data Type") +
  scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model B") +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank()) +
  theme(text=element_text(size=12))

plot_DetectionOverTime(mixLL2_list[[7]],
                              min.yr = 1900, 
                       legend_labels = c("BNM_LL1/Year effect", 
                                         "UKBMS", 
                                         "NHCs", 
                                         "BNM_LL5"), 
                       mixLL2 = TRUE,                  ### Important: specify its mixLL2
                       legend_title = "Data Type") +
  scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model C") +
  theme(text=element_text(size=12))

```



### Visits with records
```{r plot-siteprop-M-athalia, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE, fig.width=7, fig.height=4}
sppMeanSitesYear$Species <- as.factor(sppMeanSitesYear$Species)
test <- subset(sppMeanSitesYear, Species == "*Melitaea athalia*")

test$Model <- as.factor(test$Model)

require("ggtext")
ggplot(data = test,) +
  geom_ribbon(aes_string(fill = "Model", 
                    x = "Year", 
                    y = "sim_MeanSitesPerYear", 
                             ymin = "sim_lower", ymax = "sim_upper"), alpha = 0.3) +
  scale_fill_manual(values =c("#440154FF", "#20A387FF", "#E69F00")) + 
  geom_line(aes(x = Year, y = sim_MeanSitesPerYear, colour = Model), size=0.7) + 
          scale_colour_manual(values =c("#440154FF", "#20A387FF", "#E69F00")) +
  geom_line(aes(x=Year, y=obs_meanSitesPerYear), linetype = "dotted", colour=c("black"), size = 1.5) +
  geom_line(aes(x=Year, y=obs_meanSitesPerYear), colour=c("black"), size = 0.4) +
  geom_vline(xintercept = 1976, linetype = "dashed") +   # Vertical line = 1976
  ylab("Proportion of visited sites with records") + xlab("Year") +
  scale_y_continuous(limits=c(0, 0.1), expand = c(0, 0)) +
  scale_x_continuous(breaks=seq(1900, 2020, 20), 
                     limits = c(1900, 2025), 
                     expand = c(0, 0)) +
  ggtitle("Melitaea athalia") + 
  theme(plot.title=element_text(face="italic")) +
  theme_bw()
```


### Record numbers
```{r plot-record-numbers-M-athalia, message=FALSE, warning=FALSE, echo=FALSE, fig.width=7, fig.height=4, cache=TRUE}
# Subset each species' records into dataframes and specify species identity
spp_vis_M_athalia <- subset(spp_vis, Melitaea.athalia == "TRUE")
spp_vis_M_athalia$species <- c("italic('M. athalia')")

 ggplot(spp_vis_M_athalia, aes(x=year, fill=Source, color=Source)) + 
  scale_fill_manual(values = alpha(c("#453781FF", 
                                     "#20A386FF", 
                                     "#DCE318FF"), 0.5)) +
  scale_color_manual(values = alpha(c("#453781FF", 
                                      "#20A386FF", 
                                      "#DCE318FF"), 0.5)) +
  geom_bar(position="identity", alpha=0.7) + 
  labs(y="Records", x = "Year") +
  theme_bw() +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank()) +
  # sqrt y scale and expand scale to no space below, 5% above
  scale_y_sqrt(expand = expansion(mult = c(0, .05))) +
  scale_x_continuous(breaks=seq(1900, 2020, 20), 
                     limits = c(1900, 2025), 
                     expand = c(0, 0)) +
  ggtitle("Melitaea athalia") + 
  theme(plot.title=element_text(face="italic")) +
  theme(text = element_text(size=15),
        axis.text.x = element_text(angle=90, hjust=1))
```


### DIC & Bayesian *p*
```{r DIC-M-athalia, cache=TRUE, message=FALSE, echo=FALSE, cache=TRUE}
# Create empty data frame to populate
summary_out <- data.frame(Model = as.factor(c("catLL", "mixLL", "mixLL2")),
                          DIC = as.numeric(NA),
                          Bayesian_p = as.numeric(NA),
                          Bayesian_p_var = as.numeric(NA)) 


  summary_out[1:3, "DIC"] <-  c((catLL_list[[7]]$BUGSoutput$DIC), 
                                (mixLL_list[[7]]$BUGSoutput$DIC),
                                (mixLL2_list[[7]]$BUGSoutput$DIC))  
  summary_out[1, "Bayesian_p"] <-  ModelBp_Summ[7, 2]
  summary_out[2, "Bayesian_p"] <-  ModelBp_Summ[7, 3]
  summary_out[3, "Bayesian_p"] <-  ModelBp_Summ[7, 4]
  summary_out[1, "Bayesian_p_var"] <-  ModelBp_Summ[7, 5]
  summary_out[2, "Bayesian_p_var"] <-  ModelBp_Summ[7, 6]
  summary_out[3, "Bayesian_p_var"] <-  ModelBp_Summ[7, 7]

knitr::kable(summary_out)
```














4. *Polyommatus bellargus*   
5. *Thecla betulae*   
7. *Polygonia c-album*   
10. *Pararge aegeria*   
12. *Pieris rapae*   


## *Pararge aegeria* (Speckled Wood)

### Occupancy
```{r plot-occu-outputs-P-aegeria, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE}
### Source plotting function
source("./fig-plot.Occdet_G.R")

plot.occDet_G(mixLL_list[[8]], mixLL2_list[[8]], catLL_list[[8]]) +
  ggtitle(label = "Pararge aegeria", 
          subtitle = paste(lama_translate_("Pararge aegeria", dict, "common_names"))) + 
  theme(plot.title=element_text(face="italic")) +
  scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  theme(text=element_text(size=15)) 
```



### Detection probability
```{r plot-detprob-outputs-P-aegeria, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE, fig.width=7, fig.height=3}

plot_DetectionOverTime(catLL_list[[8]],
                              min.yr = 1900, 
                                    legend_labels =
                         c("    BNM/Year_Effect  ", 
                                                    "UKBMS", 
                                                   "NHCs"),
                                  legend_title = "Data Type") +
             scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
             scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model A") +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank()) +
  theme(text=element_text(size=12)) 

plot_DetectionOverTime(mixLL_list[[8]], 
                       min.yr = 1900, 
                       legend_labels = c("BNM_LL1/Year effect", 
                                         "UKBMS", 
                                         "NHCs", 
                                         "BNM_LL5"), 
                       legend_title = "Data Type") +
  scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model B") +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank()) +
  theme(text=element_text(size=12))

plot_DetectionOverTime(mixLL2_list[[8]],
                              min.yr = 1900, 
                       legend_labels = c("BNM_LL1/Year effect", 
                                         "UKBMS", 
                                         "NHCs", 
                                         "BNM_LL5"), 
                       mixLL2 = TRUE,                  ### Important: specify its mixLL2
                       legend_title = "Data Type") +
  scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model C") +
  theme(text=element_text(size=12))

```



### Visits with records
```{r plot-siteprop-P-aegeria, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE, fig.width=7, fig.height=4}
sppMeanSitesYear$Species <- as.factor(sppMeanSitesYear$Species)
test <- subset(sppMeanSitesYear, Species == "*Pararge aegeria*")

test$Model <- as.factor(test$Model)

require("ggtext")
ggplot(data = test,) +
  geom_ribbon(aes_string(fill = "Model", 
                    x = "Year", 
                    y = "sim_MeanSitesPerYear", 
                             ymin = "sim_lower", ymax = "sim_upper"), alpha = 0.3) +
  scale_fill_manual(values =c("#440154FF", "#20A387FF", "#E69F00")) + 
  geom_line(aes(x = Year, y = sim_MeanSitesPerYear, colour = Model), size=0.7) + 
          scale_colour_manual(values =c("#440154FF", "#20A387FF", "#E69F00")) +
  geom_line(aes(x=Year, y=obs_meanSitesPerYear), linetype = "dotted", colour=c("black"), size = 1.5) +
  geom_line(aes(x=Year, y=obs_meanSitesPerYear), colour=c("black"), size = 0.4) +
  geom_vline(xintercept = 1976, linetype = "dashed") +   # Vertical line = 1976
  ylab("Proportion of visited sites with records") + xlab("Year") +
  scale_y_continuous(limits=c(0, 0.8), expand = c(0, 0)) +
  scale_x_continuous(breaks=seq(1900, 2020, 20), 
                     limits = c(1900, 2025), 
                     expand = c(0, 0)) +
  ggtitle("Pararge aegeria") + 
  theme(plot.title=element_text(face="italic")) +
  theme_bw()
```


### Record numbers
```{r plot-record-numbers-P-aegeria, message=FALSE, warning=FALSE, echo=FALSE, fig.width=7, fig.height=4, cache=TRUE}
# Subset each species' records into dataframes and specify species identity
spp_vis_P_aegeria <- subset(spp_vis, Pararge.aegeria == "TRUE")
spp_vis_P_aegeria$species <- c("italic('P. aegeria')")

 ggplot(spp_vis_P_aegeria, aes(x=year, fill=Source, color=Source)) + 
  scale_fill_manual(values = alpha(c("#453781FF", 
                                     "#20A386FF", 
                                     "#DCE318FF"), 0.5)) +
  scale_color_manual(values = alpha(c("#453781FF", 
                                      "#20A386FF", 
                                      "#DCE318FF"), 0.5)) +
  geom_bar(position="identity", alpha=0.7) + 
  labs(y="Records", x = "Year") +
  theme_bw() +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank()) +
  # sqrt y scale and expand scale to no space below, 5% above
  scale_y_sqrt(expand = expansion(mult = c(0, .05))) +
  scale_x_continuous(breaks=seq(1900, 2020, 20), 
                     limits = c(1900, 2025), 
                     expand = c(0, 0)) +
  ggtitle("Pararge aegeria") + 
  theme(plot.title=element_text(face="italic")) +
  theme(text = element_text(size=15),
        axis.text.x = element_text(angle=90, hjust=1))
```


### DIC & Bayesian *p*
```{r DIC-P-aegeria, cache=TRUE, message=FALSE, echo=FALSE, cache=TRUE}
# Create empty data frame to populate
summary_out <- data.frame(Model = as.factor(c("catLL", "mixLL", "mixLL2")),
                          DIC = as.numeric(NA),
                          Bayesian_p = as.numeric(NA),
                          Bayesian_p_var = as.numeric(NA)) 


  summary_out[1:3, "DIC"] <-  c((catLL_list[[8]]$BUGSoutput$DIC), 
                                (mixLL_list[[8]]$BUGSoutput$DIC),
                                (mixLL2_list[[8]]$BUGSoutput$DIC))  
  summary_out[1, "Bayesian_p"] <-  ModelBp_Summ[8, 2]
  summary_out[2, "Bayesian_p"] <-  ModelBp_Summ[8, 3]
  summary_out[3, "Bayesian_p"] <-  ModelBp_Summ[8, 4]
  summary_out[1, "Bayesian_p_var"] <-  ModelBp_Summ[8, 5]
  summary_out[2, "Bayesian_p_var"] <-  ModelBp_Summ[8, 6]
  summary_out[3, "Bayesian_p_var"] <-  ModelBp_Summ[8, 7]

knitr::kable(summary_out)
```












   



## *Pieris rapae* (Small/Cabbage White)

### Occupancy
```{r plot-occu-outputs-P-rapae, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE}
### Source plotting function
source("./fig-plot.Occdet_G.R")

plot.occDet_G(mixLL_list[[9]], mixLL2_list[[9]], catLL_list[[9]]) +
  ggtitle(label = "Pieris rapae", 
          subtitle = paste(lama_translate_("Pieris rapae", dict, "common_names"))) + 
  theme(plot.title=element_text(face="italic")) +
  scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  theme(text=element_text(size=15)) 
```



### Detection probability
```{r plot-detprob-outputs-P-rapae, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE, fig.width=7, fig.height=3}

plot_DetectionOverTime(catLL_list[[9]],
                              min.yr = 1900, 
                                    legend_labels =
                         c("    BNM/Year_Effect  ", 
                                                    "UKBMS", 
                                                   "NHCs"),
                                  legend_title = "Data Type") +
             scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
             scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model A") +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank()) +
  theme(text=element_text(size=12)) 

plot_DetectionOverTime(mixLL_list[[9]], 
                       min.yr = 1900, 
                       legend_labels = c("BNM_LL1/Year effect", 
                                         "UKBMS", 
                                         "NHCs", 
                                         "BNM_LL5"), 
                       legend_title = "Data Type") +
  scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model B") +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank()) +
  theme(text=element_text(size=12))

plot_DetectionOverTime(mixLL2_list[[9]],
                              min.yr = 1900, 
                       legend_labels = c("BNM_LL1/Year effect", 
                                         "UKBMS", 
                                         "NHCs", 
                                         "BNM_LL5"), 
                       mixLL2 = TRUE,                  ### Important: specify its mixLL2
                       legend_title = "Data Type") +
  scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model C") +
  theme(text=element_text(size=12))

```



### Visits with records
```{r plot-siteprop-P-rapae, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE, fig.width=7, fig.height=4}
sppMeanSitesYear$Species <- as.factor(sppMeanSitesYear$Species)
test <- subset(sppMeanSitesYear, Species == "*Pieris rapae*")

test$Model <- as.factor(test$Model)

require("ggtext")
ggplot(data = test,) +
  geom_ribbon(aes_string(fill = "Model", 
                    x = "Year", 
                    y = "sim_MeanSitesPerYear", 
                             ymin = "sim_lower", ymax = "sim_upper"), alpha = 0.3) +
  scale_fill_manual(values =c("#440154FF", "#20A387FF", "#E69F00")) + 
  geom_line(aes(x = Year, y = sim_MeanSitesPerYear, colour = Model), size=0.7) + 
          scale_colour_manual(values =c("#440154FF", "#20A387FF", "#E69F00")) +
  geom_line(aes(x=Year, y=obs_meanSitesPerYear), linetype = "dotted", colour=c("black"), size = 1.5) +
  geom_line(aes(x=Year, y=obs_meanSitesPerYear), colour=c("black"), size = 0.4) +
  geom_vline(xintercept = 1976, linetype = "dashed") +   # Vertical line = 1976
  ylab("Proportion of visited sites with records") + xlab("Year") +
  scale_y_continuous(limits=c(0, 0.8), expand = c(0, 0)) +
  scale_x_continuous(breaks=seq(1900, 2020, 20), 
                     limits = c(1900, 2025), 
                     expand = c(0, 0)) +
  ggtitle("Pieris rapae") + 
  theme(plot.title=element_text(face="italic")) +
  theme_bw()
```


### Record numbers
```{r plot-record-numbers-P-rapae, message=FALSE, warning=FALSE, echo=FALSE, fig.width=7, fig.height=4, cache=TRUE}
# Subset each species' records into dataframes and specify species identity
spp_vis_P_rapae <- subset(spp_vis, Pieris.rapae == "TRUE")
spp_vis_P_rapae$species <- c("italic('P. rapae')")

 ggplot(spp_vis_P_rapae, aes(x=year, fill=Source, color=Source)) + 
  scale_fill_manual(values = alpha(c("#453781FF", 
                                     "#20A386FF", 
                                     "#DCE318FF"), 0.5)) +
  scale_color_manual(values = alpha(c("#453781FF", 
                                      "#20A386FF", 
                                      "#DCE318FF"), 0.5)) +
  geom_bar(position="identity", alpha=0.7) + 
  labs(y="Records", x = "Year") +
  theme_bw() +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank()) +
  # sqrt y scale and expand scale to no space below, 5% above
  scale_y_sqrt(expand = expansion(mult = c(0, .05))) +
  scale_x_continuous(breaks=seq(1900, 2020, 20), 
                     limits = c(1900, 2025), 
                     expand = c(0, 0)) +
  ggtitle("Pieris rapae") + 
  theme(plot.title=element_text(face="italic")) +
  theme(text = element_text(size=15),
        axis.text.x = element_text(angle=90, hjust=1))
```


### DIC & Bayesian *p*
```{r DIC-P-rapae, cache=TRUE, message=FALSE, echo=FALSE, cache=TRUE}
# Create empty data frame to populate
summary_out <- data.frame(Model = as.factor(c("catLL", "mixLL", "mixLL2")),
                          DIC = as.numeric(NA),
                          Bayesian_p = as.numeric(NA),
                          Bayesian_p_var = as.numeric(NA)) 


  summary_out[1:3, "DIC"] <-  c((catLL_list[[9]]$BUGSoutput$DIC), 
                                (mixLL_list[[9]]$BUGSoutput$DIC),
                                (mixLL2_list[[9]]$BUGSoutput$DIC))  
  summary_out[1, "Bayesian_p"] <-  ModelBp_Summ[9, 2]
  summary_out[2, "Bayesian_p"] <-  ModelBp_Summ[9, 3]
  summary_out[3, "Bayesian_p"] <-  ModelBp_Summ[9, 4]
  summary_out[1, "Bayesian_p_var"] <-  ModelBp_Summ[9, 5]
  summary_out[2, "Bayesian_p_var"] <-  ModelBp_Summ[9, 6]
  summary_out[3, "Bayesian_p_var"] <-  ModelBp_Summ[9, 7]

knitr::kable(summary_out)
```




















## *Polygonia c-album* (Comma)

### Occupancy
```{r plot-occu-outputs-Polygonia-c-album, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE}
### Source plotting function
source("./fig-plot.Occdet_G.R")

plot.occDet_G(mixLL_list[[10]], mixLL2_list[[10]], catLL_list[[10]]) +
  ggtitle(label = "Polygonia c-album", 
          subtitle = paste(lama_translate_("Polygonia c-album", dict, "common_names"))) + 
  theme(plot.title=element_text(face="italic")) +
  scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  theme(text=element_text(size=15)) 
```



### Detection probability
```{r plot-detprob-outputs-P-c-album, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE, fig.width=7, fig.height=3}
### Source plotting function
source("/Users/galinajonsson/Documents/PhD_work/Collaborations/sparta/sparta/R/plot_DetectionOverTime.R")


plot_DetectionOverTime(catLL_list[[10]],
                              min.yr = 1900, 
                                    legend_labels =
                         c("    BNM/Year_Effect  ", 
                                                    "UKBMS", 
                                                   "NHCs"),
                                  legend_title = "Data Type") +
             scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
             scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model A") +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank()) +
  theme(text=element_text(size=12)) 

plot_DetectionOverTime(mixLL_list[[10]], 
                       min.yr = 1900, 
                       legend_labels = c("BNM_LL1/Year effect", 
                                         "UKBMS", 
                                         "NHCs", 
                                         "BNM_LL5"), 
                       legend_title = "Data Type") +
  scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model B") +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank()) +
  theme(text=element_text(size=12))

plot_DetectionOverTime(mixLL2_list[[10]],
                              min.yr = 1900, 
                       legend_labels = c("BNM_LL1/Year effect", 
                                         "UKBMS", 
                                         "NHCs", 
                                         "BNM_LL5"), 
                       mixLL2 = TRUE,                  ### Important: specify its mixLL2
                       legend_title = "Data Type") +
  scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model C") +
  theme(text=element_text(size=12))

```



### Visits with records
```{r plot-siteprop-P-c-album, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE, fig.width=7, fig.height=4}
sppMeanSitesYear$Species <- as.factor(sppMeanSitesYear$Species)
test <- subset(sppMeanSitesYear, Species == "*Polygonia c-album*")

test$Model <- as.factor(test$Model)

require("ggtext")
ggplot(data = test,) +
  geom_ribbon(aes_string(fill = "Model", 
                    x = "Year", 
                    y = "sim_MeanSitesPerYear", 
                             ymin = "sim_lower", ymax = "sim_upper"), alpha = 0.3) +
  scale_fill_manual(values =c("#440154FF", "#20A387FF", "#E69F00")) + 
  geom_line(aes(x = Year, y = sim_MeanSitesPerYear, colour = Model), size=0.7) + 
          scale_colour_manual(values =c("#440154FF", "#20A387FF", "#E69F00")) +
  geom_line(aes(x=Year, y=obs_meanSitesPerYear), linetype = "dotted", colour=c("black"), size = 1.5) +
  geom_line(aes(x=Year, y=obs_meanSitesPerYear), colour=c("black"), size = 0.4) +
  geom_vline(xintercept = 1976, linetype = "dashed") +   # Vertical line = 1976
  ylab("Proportion of visited sites with records") + xlab("Year") +
  scale_y_continuous(limits=c(0, 0.8), expand = c(0, 0)) +
  scale_x_continuous(breaks=seq(1900, 2020, 20), 
                     limits = c(1900, 2025), 
                     expand = c(0, 0)) +
  ggtitle("Polygonia c-album") + 
  theme(plot.title=element_text(face="italic")) +
  theme_bw()
```


### Record numbers
```{r plot-record-numbers-P-c-album, message=FALSE, warning=FALSE, echo=FALSE, fig.width=7, fig.height=4, cache=TRUE}
# Subset each species' records into dataframes and specify species identity
spp_vis_P_c_album <- subset(spp_vis, Polygonia.c.album == "TRUE")
spp_vis_P_c_album$species <- c("italic('P. c-album')")

 ggplot(spp_vis_P_c_album, aes(x=year, fill=Source, color=Source)) + 
  scale_fill_manual(values = alpha(c("#453781FF", 
                                     "#20A386FF", 
                                     "#DCE318FF"), 0.5)) +
  scale_color_manual(values = alpha(c("#453781FF", 
                                      "#20A386FF", 
                                      "#DCE318FF"), 0.5)) +
  geom_bar(position="identity", alpha=0.7) + 
  labs(y="Records", x = "Year") +
  theme_bw() +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank()) +
  # sqrt y scale and expand scale to no space below, 5% above
  scale_y_sqrt(expand = expansion(mult = c(0, .05))) +
  scale_x_continuous(breaks=seq(1900, 2020, 20), 
                     limits = c(1900, 2025), 
                     expand = c(0, 0)) +
  ggtitle("Polygonia c-album") + 
  theme(plot.title=element_text(face="italic")) +
  theme(text = element_text(size=15),
        axis.text.x = element_text(angle=90, hjust=1))
```


### DIC & Bayesian *p*
```{r DIC-P-c-album, cache=TRUE, message=FALSE, echo=FALSE, cache=TRUE}
# Create empty data frame to populate
summary_out <- data.frame(Model = as.factor(c("catLL", "mixLL", "mixLL2")),
                          DIC = as.numeric(NA),
                          Bayesian_p = as.numeric(NA),
                          Bayesian_p_var = as.numeric(NA)) 


  summary_out[1:3, "DIC"] <-  c((catLL_list[[10]]$BUGSoutput$DIC), 
                                (mixLL_list[[10]]$BUGSoutput$DIC),
                                (mixLL2_list[[10]]$BUGSoutput$DIC))  
  summary_out[1, "Bayesian_p"] <-  ModelBp_Summ[10, 2]
  summary_out[2, "Bayesian_p"] <-  ModelBp_Summ[10, 3]
  summary_out[3, "Bayesian_p"] <-  ModelBp_Summ[10, 4]
  summary_out[1, "Bayesian_p_var"] <-  ModelBp_Summ[10, 5]
  summary_out[2, "Bayesian_p_var"] <-  ModelBp_Summ[10, 6]
  summary_out[3, "Bayesian_p_var"] <-  ModelBp_Summ[10, 7]

knitr::kable(summary_out)
```













4. *Polyommatus bellargus*   
5. *Thecla betulae*   






## *Polyommatus bellargus* (Adonis Blue)

### Occupancy
```{r plot-occu-outputs-P-bellargus, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE}
### Source plotting function
source("./fig-plot.Occdet_G.R")

plot.occDet_G(mixLL_list[[11]], mixLL2_list[[11]], catLL_list[[11]]) +
  ggtitle(label = "Polyommatus bellargus", 
          subtitle = paste(lama_translate_("Polyommatus bellargus", dict, "common_names"))) + 
  theme(plot.title=element_text(face="italic")) +
  scale_y_continuous(limits=c(0, 0.1), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  theme(text=element_text(size=15)) 
```



### Detection probability
```{r plot-detprob-outputs-P-bellargus, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE, fig.width=7, fig.height=3}

plot_DetectionOverTime(catLL_list[[11]],
                              min.yr = 1900, 
                                    legend_labels =
                         c("    BNM/Year_Effect  ", 
                                                    "UKBMS", 
                                                   "NHCs"),
                                  legend_title = "Data Type") +
             scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
             scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model A") +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank()) +
  theme(text=element_text(size=12)) 

plot_DetectionOverTime(mixLL_list[[11]], 
                       min.yr = 1900, 
                       legend_labels = c("BNM_LL1/Year effect", 
                                         "UKBMS", 
                                         "NHCs", 
                                         "BNM_LL5"), 
                       legend_title = "Data Type") +
  scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model B") +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank()) +
  theme(text=element_text(size=12))

plot_DetectionOverTime(mixLL2_list[[11]],
                              min.yr = 1900, 
                       legend_labels = c("BNM_LL1/Year effect", 
                                         "UKBMS", 
                                         "NHCs", 
                                         "BNM_LL5"), 
                       mixLL2 = TRUE,                  ### Important: specify its mixLL2
                       legend_title = "Data Type") +
  scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model C") +
  theme(text=element_text(size=12))

```



### Visits with records
```{r plot-siteprop-P-bellargus, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE, fig.width=7, fig.height=4}
sppMeanSitesYear$Species <- as.factor(sppMeanSitesYear$Species)
test <- subset(sppMeanSitesYear, Species == "*Polyommatus bellargus*")

test$Model <- as.factor(test$Model)

require("ggtext")
ggplot(data = test,) +
  geom_ribbon(aes_string(fill = "Model", 
                    x = "Year", 
                    y = "sim_MeanSitesPerYear", 
                             ymin = "sim_lower", ymax = "sim_upper"), alpha = 0.3) +
  scale_fill_manual(values =c("#440154FF", "#20A387FF", "#E69F00")) + 
  geom_line(aes(x = Year, y = sim_MeanSitesPerYear, colour = Model), size=0.7) + 
          scale_colour_manual(values =c("#440154FF", "#20A387FF", "#E69F00")) +
  geom_line(aes(x=Year, y=obs_meanSitesPerYear), linetype = "dotted", colour=c("black"), size = 1.5) +
  geom_line(aes(x=Year, y=obs_meanSitesPerYear), colour=c("black"), size = 0.4) +
  geom_vline(xintercept = 1976, linetype = "dashed") +   # Vertical line = 1976
  ylab("Proportion of visited sites with records") + xlab("Year") +
  scale_y_continuous(limits=c(0, 0.2), expand = c(0, 0)) +
  scale_x_continuous(breaks=seq(1900, 2020, 20), 
                     limits = c(1900, 2025), 
                     expand = c(0, 0)) +
  ggtitle("Polyommatus bellargus") + 
  theme(plot.title=element_text(face="italic")) +
  theme_bw()
```


### Record numbers
```{r plot-record-numbers-P-bellargus, message=FALSE, warning=FALSE, echo=FALSE, fig.width=7, fig.height=4, cache=TRUE}
# Subset each species' records into dataframes and specify species identity
spp_vis_P_bellargus <- subset(spp_vis, Polyommatus.bellargus == "TRUE")
spp_vis_P_bellargus$species <- c("italic('P. bellargus')")

 ggplot(spp_vis_P_bellargus, aes(x=year, fill=Source, color=Source)) + 
  scale_fill_manual(values = alpha(c("#453781FF", 
                                     "#20A386FF", 
                                     "#DCE318FF"), 0.5)) +
  scale_color_manual(values = alpha(c("#453781FF", 
                                      "#20A386FF", 
                                      "#DCE318FF"), 0.5)) +
  geom_bar(position="identity", alpha=0.7) + 
  labs(y="Records", x = "Year") +
  theme_bw() +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank()) +
  # sqrt y scale and expand scale to no space below, 5% above
  scale_y_sqrt(expand = expansion(mult = c(0, .05))) +
  scale_x_continuous(breaks=seq(1900, 2020, 20), 
                     limits = c(1900, 2025), 
                     expand = c(0, 0)) +
  ggtitle("Polyommatus bellargus") + 
  theme(plot.title=element_text(face="italic")) +
  theme(text = element_text(size=15),
        axis.text.x = element_text(angle=90, hjust=1))
```


### DIC & Bayesian *p*
```{r DIC-P-bellargus, cache=TRUE, message=FALSE, echo=FALSE, cache=TRUE}
# Create empty data frame to populate
summary_out <- data.frame(Model = as.factor(c("catLL", "mixLL", "mixLL2")),
                          DIC = as.numeric(NA),
                          Bayesian_p = as.numeric(NA),
                          Bayesian_p_var = as.numeric(NA)) 


  summary_out[1:3, "DIC"] <-  c((catLL_list[[11]]$BUGSoutput$DIC), 
                                (mixLL_list[[11]]$BUGSoutput$DIC),
                                (mixLL2_list[[11]]$BUGSoutput$DIC))  
  summary_out[1, "Bayesian_p"] <-  ModelBp_Summ[11, 2]
  summary_out[2, "Bayesian_p"] <-  ModelBp_Summ[11, 3]
  summary_out[3, "Bayesian_p"] <-  ModelBp_Summ[11, 4]
  summary_out[1, "Bayesian_p_var"] <-  ModelBp_Summ[11, 5]
  summary_out[2, "Bayesian_p_var"] <-  ModelBp_Summ[11, 6]
  summary_out[3, "Bayesian_p_var"] <-  ModelBp_Summ[11, 7]

knitr::kable(summary_out)
```





 




## *Thecla betulae* (Brown Hairstreak)

### Occupancy
```{r plot-occu-outputs-P-betulae, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE}
### Source plotting function
source("./fig-plot.Occdet_G.R")

plot.occDet_G(mixLL_list[[12]], mixLL2_list[[12]], catLL_list[[12]]) +
  ggtitle(label = "Thecla betulae", 
          subtitle = paste(lama_translate_("Thecla betulae", dict, "common_names"))) + 
  theme(plot.title=element_text(face="italic")) +
  scale_y_continuous(limits=c(0, 0.5), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  theme(text=element_text(size=15)) 
```



### Detection probability
```{r plot-detprob-outputs-P-betulae, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE, fig.width=7, fig.height=3}

plot_DetectionOverTime(catLL_list[[12]],
                              min.yr = 1900, 
                                    legend_labels =
                         c("    BNM/Year_Effect  ", 
                                                    "UKBMS", 
                                                   "NHCs"),
                                  legend_title = "Data Type") +
             scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
             scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model A") +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank()) +
  theme(text=element_text(size=12)) 

plot_DetectionOverTime(mixLL_list[[12]], 
                       min.yr = 1900, 
                       legend_labels = c("BNM_LL1/Year effect", 
                                         "UKBMS", 
                                         "NHCs", 
                                         "BNM_LL5"), 
                       legend_title = "Data Type") +
  scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model B") +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank()) +
  theme(text=element_text(size=12))

plot_DetectionOverTime(mixLL2_list[[12]],
                              min.yr = 1900, 
                       legend_labels = c("BNM_LL1/Year effect", 
                                         "UKBMS", 
                                         "NHCs", 
                                         "BNM_LL5"), 
                       mixLL2 = TRUE,                  ### Important: specify its mixLL2
                       legend_title = "Data Type") +
  scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  ggtitle("Model C") +
  theme(text=element_text(size=12))

```



### Visits with records
```{r plot-siteprop-P-betulae, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE, fig.width=7, fig.height=4}
sppMeanSitesYear$Species <- as.factor(sppMeanSitesYear$Species)
test <- subset(sppMeanSitesYear, Species == "*Thecla betulae*")

test$Model <- as.factor(test$Model)

require("ggtext")
ggplot(data = test,) +
  geom_ribbon(aes_string(fill = "Model", 
                    x = "Year", 
                    y = "sim_MeanSitesPerYear", 
                             ymin = "sim_lower", ymax = "sim_upper"), alpha = 0.3) +
  scale_fill_manual(values =c("#440154FF", "#20A387FF", "#E69F00")) + 
  geom_line(aes(x = Year, y = sim_MeanSitesPerYear, colour = Model), size=0.7) + 
          scale_colour_manual(values =c("#440154FF", "#20A387FF", "#E69F00")) +
  geom_line(aes(x=Year, y=obs_meanSitesPerYear), linetype = "dotted", colour=c("black"), size = 1.5) +
  geom_line(aes(x=Year, y=obs_meanSitesPerYear), colour=c("black"), size = 0.4) +
  geom_vline(xintercept = 1976, linetype = "dashed") +   # Vertical line = 1976
  ylab("Proportion of visited sites with records") + xlab("Year") +
  scale_y_continuous(limits=c(0, 0.1), expand = c(0, 0)) +
  scale_x_continuous(breaks=seq(1900, 2020, 20), 
                     limits = c(1900, 2025), 
                     expand = c(0, 0)) +
  ggtitle("Thecla betulae") + 
  theme(plot.title=element_text(face="italic")) +
  theme_bw()
```


### Record numbers
```{r plot-record-numbers-P-betulae, message=FALSE, warning=FALSE, echo=FALSE, fig.width=7, fig.height=4, cache=TRUE}
# Subset each species' records into dataframes and specify species identity
spp_vis_T_betulae <- subset(spp_vis, Thecla.betulae == "TRUE")
spp_vis_T_betulae$species <- c("italic('T. betulae')")

 ggplot(spp_vis_T_betulae, aes(x=year, fill=Source, color=Source)) + 
  scale_fill_manual(values = alpha(c("#453781FF", 
                                     "#20A386FF", 
                                     "#DCE318FF"), 0.5)) +
  scale_color_manual(values = alpha(c("#453781FF", 
                                      "#20A386FF", 
                                      "#DCE318FF"), 0.5)) +
  geom_bar(position="identity", alpha=0.7) + 
  labs(y="Records", x = "Year") +
  theme_bw() +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank()) +
  # sqrt y scale and expand scale to no space below, 5% above
  scale_y_sqrt(expand = expansion(mult = c(0, .05))) +
  scale_x_continuous(breaks=seq(1900, 2020, 20), 
                     limits = c(1900, 2025), 
                     expand = c(0, 0)) +
  ggtitle("Thecla betulae") + 
  theme(plot.title=element_text(face="italic")) +
  theme(text = element_text(size=15),
        axis.text.x = element_text(angle=90, hjust=1))
```


### DIC & Bayesian *p*
```{r DIC-P-betulae, cache=TRUE, message=FALSE, echo=FALSE, cache=TRUE}
# Create empty data frame to populate
summary_out <- data.frame(Model = as.factor(c("catLL", "mixLL", "mixLL2")),
                          DIC = as.numeric(NA),
                          Bayesian_p = as.numeric(NA),
                          Bayesian_p_var = as.numeric(NA)) 


  summary_out[1:3, "DIC"] <-  c((catLL_list[[12]]$BUGSoutput$DIC), 
                                (mixLL_list[[12]]$BUGSoutput$DIC),
                                (mixLL2_list[[12]]$BUGSoutput$DIC))  
  summary_out[1, "Bayesian_p"] <-  ModelBp_Summ[12, 2]
  summary_out[2, "Bayesian_p"] <-  ModelBp_Summ[12, 3]
  summary_out[3, "Bayesian_p"] <-  ModelBp_Summ[12, 4]
  summary_out[1, "Bayesian_p_var"] <-  ModelBp_Summ[12, 5]
  summary_out[2, "Bayesian_p_var"] <-  ModelBp_Summ[12, 6]
  summary_out[3, "Bayesian_p_var"] <-  ModelBp_Summ[12, 7]

knitr::kable(summary_out)
```
















###### Shiny plots - not used but to annotate
```{r SHINY-plot-occu-outputs, message=FALSE, warning=FALSE, echo=FALSE, eval=FALSE}
require(shiny)
require(ggplot2)
require(labelmachine)
require(sparta)
require(dplyr)

### Source plotting function
source("./fig-plot.Occdet_G.R")

# Change ModelNames vetor to species names with spaces
spp_names <- gsub(ModelNames, patt = "_", repl=" ")
spp_names <- sub(spp_names, patt = "c.album", repl="c-album")

# Name the lists of outputs their species names
names(mixLL_list)[1:12] <- spp_names
names(mixLL2_list)[1:12] <- spp_names


# Create a dictionary translating Latin names to common names
dict <- new_lama_dictionary(common_names = 
                              c("Aglais urticae" = 'Small Tortoiseshell',
                                "Carterocephalus palaemon" = 'Chequered Skipper',
                                "Erebia aethiops" = 'Scotch Argus',
                                "Erebia epiphron" = 'Mountain Ringlet',
                                "Limenitis camilla" = 'White Admiral',
                                "Lycaena phlaeas" = 'Small Copper',
                                "Melitaea athalia" = 'Heath Fritillary',
                                "Pararge aegeria" = 'Speckled Wood',
                                "Pieris rapae" = 'Small White',
                                "Polygonia c-album" = 'Comma',
                                "Polyommatus bellargus" = 'Adonis Blue',
                                "Thecla betulae" = 'Brown Hairstreak'))


# Plot occupancy estimates as a shiny app (currently two models per species)

shinyApp(
  ui = fluidPage(
    # Select input
    selectInput("species_choice", # call the variable storing values (spp names) "species_choice"
                label = "Select species", # header text of dropdown list
                choices = spp_names, # vector of all values in dropdown list—in (species)
                selected ="Aglais urticae"), # dropdown defaults to Aglais urticae
    plotOutput("occuPlot")
  ),
  server = function(input, output) {
    output$occuPlot <- renderPlot({
      
      vec <- paste(input$species_choice)  
      vec_labeled <- lama_translate_(vec, dict, "common_names")  
      
      plot.occDet_G(mixLL_list[[input$species_choice]], mixLL2_list[[input$species_choice]]) +
        ggtitle(label = paste(input$species_choice), 
                subtitle = paste(vec_labeled)) + 
        theme(plot.title=element_text(face="italic")) +
        theme(text=element_text(size=18))
    })
  },
  options = list(height = 500)
)
```


 
```{r SHINY-plot-det-outputs, message=FALSE, echo=FALSE, fig.show='hold', out.width='50%', eval=FALSE}

require(shiny)
require(ggplot2)
require(sparta)
require(dplyr)

 
# Plot detection probability estimates as a shiny app (currently two models per species)
 
selectInput("species_choice", 
            label = "Select species",
            choices = spp_names, 
            selected ="Aglais urticae")
fluidPage(fluidRow(
#  column(5, 
#         renderPlot(
#           plot_DetectionOverTime(catLL_list[[input$species_choice]], 
#                                  min.yr = 1900) 
#             theme(legend.position = "none") +
#             scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
#             scale_x_continuous(expand = c(0, 0)) +
#             ggtitle(label = paste(input$species_choice), 
#                subtitle = "Model A") + 
#             theme(plot.title=element_text(face="italic")) +
#             theme(text=element_text(size=16))
#         )
#         ),
  column(5, 
         renderPlot(
           plot_DetectionOverTime(mixLL_list[[input$species_choice]], 
                                  min.yr = 1900) + 
             theme(legend.position = "none") +
             scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
             scale_x_continuous(expand = c(0, 0)) +
             ggtitle(label = paste(input$species_choice), 
                     subtitle = "Model B") + 
             theme(plot.title=element_text(face="italic")) +
             theme(text=element_text(size=16))
         )
         ),
  column(7, 
         renderPlot(
           plot_DetectionOverTime(mixLL2_list[[input$species_choice]], 
                                  min.yr = 1900, 
                                  legend_labels = c("BNM_LL1/Year_Effect", 
                                                    "UKBMS", 
                                                    "NHCs", 
                                                    "BNM_LL5"), 
                                  legend_title = "Data Type") +
             scale_y_continuous(limits=c(0, 1), expand = c(0, 0)) +
             scale_x_continuous(expand = c(0, 0)) +
             ggtitle(label = paste(input$species_choice, 
                subtitle = "Model C") + 
             theme(axis.title.y = element_blank(),
                   axis.text.y = element_blank(),
                   text=element_text(size=16))
           )
         ) 
  )
)
 
```
