==> analyses/07-SiteYearHandover.Rmd <==
---
title: "9-SiteYearHandover"
author: "Galina M. Jönsson"
date: "15/09/2021"
output: html_document
---

## All sites and site:year combinations
```{r overlapping-records, message=FALSE, warning=FALSE, echo=FALSE, cache=TRUE}
occDetdata <- read.csv("../data/formattedData/occDetdata_merged_2020-04-07.csv", 
                       header=T, na.strings=c("","NA")) 
spp_vis <- read.csv("../data/formattedData/spp_vis_merged_2020-04-07.csv", 
                    header=T, na.strings=c("","NA"))


taxa_names <- c('Aglais.urticae',
                'Carterocephalus.palaemon',
                'Erebia.aethiops',
                'Erebia.epiphron',
                'Limenitis.camilla',
                'Lycaena.phlaeas',
                'Melitaea.athalia',
                'Pararge.aegeria',
                'Pieris.rapae',
                'Polygonia.c.album',
                'Polyommatus.bellargus',
                'Thecla.betulae') 


source("./function-DT.overlap.R")

interval <- c(1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16)
for(taxon in taxa_names){
out <- DToverlap(taxon_name = paste0(taxa_names[[1]]), #paste0(taxon),
                         occDetdata = occDetdata,
                         spp_vis = spp_vis,
                         modeltype = "mixlistlength",
                         minTP = 61, 
                         maxTP = 76,
                         TPinterval = interval)


temp <- as.data.frame(matrix(ncol=7,nrow=3*16))
names(temp) <- c("TPbin", "DTcomp", "sites", "site_TPs", "TPs", "site_TPs_total", "site_TPs_species")
temp$TPbin <-  rep(interval, each= 3)
temp$DTcomp <-  rep(c("1-2", "1-3", "2-3"), times= 16)
      
count <- 1
for (i in interval) {
  temp[count:(count+2), 3:5] <- out[[i]]$overlap[1:3,1:3]
  temp[count:(count+2), 6] <- rep(out[[i]]$data[1,5], 3)
  temp[count:(count+2), 7] <- rep(out[[i]]$data[5,5], 3)
  count <- count+3
}

temp <- subset(temp, DTcomp == "1-3")

# Plot
title <- paste("Overlapping BNM-NHC records 1960-1975 \n ", paste0(taxon), sep="")
plot(temp$TPbin, temp$site_TPs,
     col = "black",
     main = title,
     xlab = "Year bins",
     ylab= "Site:Time Period Combinations")
points(temp$TPbin, temp$site_TPs,
     col = "black",)
}
```

```{r est-overlapping-records, message=FALSE, warning=FALSE, echo=FALSE, eval=FALSE}

taxa_names <- c('Aglais.urticae',
                'Carterocephalus.palaemon',
                'Erebia.aethiops',
                'Erebia.epiphron',
                'Limenitis.camilla',
                'Lycaena.phlaeas',
                'Melitaea.athalia',
                'Pararge.aegeria',
                'Pieris.rapae',
                'Polygonia.c.album',
                'Polyommatus.bellargus',
                'Thecla.betulae') 


source("./function-DT.overlap.R")

interval <- c(1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16)
test <- DToverlap(taxon_name = 'Melitaea.athalia',
                         occDetdata = occDetdata,
                         spp_vis = spp_vis,
                         modeltype = "mixlistlength",
                         minTP = 61, 
                         maxTP = 76,
                         TPinterval = interval)

```
## Aglais urticae
```{r A-urticae, message=FALSE, warning=FALSE, echo=FALSE, eval=TRUE}
require(sparta)
results_Aglais_urticae_catLL_15yrTP <- readRDS("../outputs/results_Aglais_urticae_catLL_15yrTP_3prior.rds")
results_Aglais_urticae_catLL_10yrTP <- readRDS("../outputs/results_Aglais_urticae_catLL_10yrTP_3prior.rds")
results_Aglais_urticae_catLL_5yrTP <- readRDS("../outputs/results_Aglais_urticae_catLL_5yrTP_3prior.rds")
results_Aglais_urticae_catLL_1yrTP <- readRDS("../outputs/results_Aglais_urticae_prior-3_catLL2_7000iter.rds")
Aglais_urticae_watson_catLL <- readRDS("../outputs/catLL-outputs/results_Aglais_urticae_watson_catLL.rds")
plot(Aglais_urticae_watson_catLL, main = "old prior, 1 year")
plot(results_Aglais_urticae_catLL_1yrTP, main = "new prior, 1 year")
plot(results_Aglais_urticae_catLL_5yrTP, main = "new prior, 5 years")
plot(results_Aglais_urticae_catLL_10yrTP, main = "new prior, 10 years")
plot(results_Aglais_urticae_catLL_15yrTP, main = "new prior, 15 years")
```
